% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group.R
\name{add_group}
\alias{add_group}
\title{Stratify population into groups}
\usage{
add_group(
  peterlist,
  groupnames,
  grouptype = NA,
  combinewithgrouptype = "",
  basestates = "",
  interactionscale = 1,
  transitionscale = 1
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{groupnames}{character vector of groups
in a given grouptype (e.g., c("Old","Young"))}

\item{grouptype}{character vector of grouptype (e.g., c("Age"))
default is NA which automatically generates a type during
model compilation}

\item{combinewithgrouptype}{character vector
specifying which group types to create combinations
of state names. default is "" which stratifies given
grouptype with all other grouptypes}

\item{basestates}{character vector of basestates
to stratify. default is "" which specifies all basestates}

\item{interactionscale}{character/numeric specifies
how to scale interaction rates for each group. default is 1}

\item{transitionscale}{character/numeric specifies
how to scale transition rates for each group. default is 1.}
}
\value{
updated model instructions
}
\description{
Stratify population into groups
}
