% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grouptransition.R
\name{add_grouptransition}
\alias{add_grouptransition}
\title{add transition between groups}
\usage{
add_grouptransition(
  peterlist,
  fromgroup,
  togroups,
  meantime = NA,
  rate = NA,
  processname = NA,
  processgroup = NA,
  chainlength = 1,
  chaintimescale = NA,
  forkprobability = NA,
  percapita = TRUE,
  fromstate = "",
  tostates = "",
  metapopulation = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{fromgroup}{character specifying group name to transition from}

\item{togroups}{character vector specifying named group(s) to transition to}

\item{meantime}{character/numeric value specifying
average time to go from fromstate to all tostates}

\item{rate}{character/numeric value specifying rate to go from fromstate to
all tostates}

\item{processname}{character value for referencing the added process
by other functions
default to NA}

\item{processgroup}{character value that groups named process to quickly
reference multiple processes by other functions
default to NA}

\item{chainlength}{total steps from fromstate to
tostates (i.e., boxcars or length of pitchfork)
Alters distribution of transition times during
stochastic implementation
default to 1 which is exponentially
distrubuted transition times}

\item{chaintimescale}{character/numeric vector
with length chainlength scales rate transitioning
between steps along chains. Allows deviating
transition time distribution from a gamma distributions
default to NA which is same transition times between
all chained states (e.g., gamma distributed wait times)}

\item{forkprobability}{character/numeric vector
specifying relative weight to transtion to
different tostates when multiple are given
default to NA which specifies equal probability
transitioning to each final state}

\item{percapita}{specifies rates are given as per
capita value so the total change to population
should be scaled by the tostate
default is TRUE}

\item{fromstate}{basestate to transition from}

\item{tostates}{character vector of basestates to
transition to}

\item{metapopulation}{character vector of metapopulation
names this transitions occurs in.
default "" is which specifies all metapopulations}
}
\value{
updated instruction list
}
\description{
Add transitions between basestates AND specified groups, fixing all other
state features (e.g., other groups, metapopulations). Assumes
a "pitchfork" shape where 1 state transitions to one of multiple states after
one or multiple steps.
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
