% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interaction.R
\name{add_interaction}
\alias{add_interaction}
\title{add interactions to model}
\usage{
add_interaction(
  peterlist,
  states_in,
  states_out,
  rate,
  normlogic = TRUE,
  metapopulation = "",
  processname = NA,
  processgroup = NA,
  groupnames = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{states_in}{character vector of base state names
that interact}

\item{states_out}{character vector of base
state names after interactions. Assume states_in
and states_out are ordered (e.g., S,I -> I,I)}

\item{rate}{character/numeric value specifying
PER CAPITA rate of interaction}

\item{normlogic}{logical specifying if rate
should be divided by total population in given
metapopulation}

\item{metapopulation}{character vector of
metapopulations where this interaction occurs
default is "" which specifies all metapopulations}

\item{processname}{character value for referencing the added process
by other functions
default to NA}

\item{processgroup}{character value that groups named process to quickly
reference multiple processes by other functions
default to NA}

\item{groupnames}{Groups the transition applies to.
Either a character vector of groupnames or named list with grouptype names
and groupname values
default is "" which specifies all groups.}
}
\value{
updated instruction list
}
\description{
general form where users can specify states up
and down. Assumes per capita rates of interaction!
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
