% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_replacement.R
\name{add_replacement}
\alias{add_replacement}
\title{add replacement in model}
\usage{
add_replacement(
  peterlist,
  statesout,
  statesin,
  meantime,
  rate,
  processname = "replacement",
  processgroup = NA,
  chainlength = 1,
  chaintimescale = NA,
  percapita = TRUE,
  metapopulation = "",
  groupname = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{statesout}{character vector of state names
to transition from}

\item{statesin}{character vector of basestates to
transition to}

\item{meantime}{character/numeric value specifying
average time to go from statesout to statesin}

\item{rate}{character/numeric value specifying rate to go from statesout to
statesin}

\item{processname}{character value for referencing the added process
by other functions
default is "replacement"}

\item{processgroup}{character value that groups named process to quickly
reference multiple processes by other functions
default to NA}

\item{chainlength}{total steps from statesout to
statesin (i.e., boxcars or length of pitchfork)
Alters distribution of transition times during
stochastic implementation
default to 1 which gives exponentially distributed transition times}

\item{chaintimescale}{character/numeric vector
with length chainlength scales rate transitioning
between steps along chains. Allows deviating
transition time distribution from a gamma distributions
default to NA which is same transition times between
all chained states (e.g., gamma distributed wait times)}

\item{percapita}{specifies rates are given as per
capita value so the total change to population
should be scaled by the tostate
default is TRUE}

\item{metapopulation}{character vector of metapopulation
names this transitions occurs in.
default "" is which specifies all metapopulations}

\item{groupname}{Groups the transition applies to.
Either a character vector of groupnames or named list with grouptype names
and groupname values
default is "" which specifies all groups.}
}
\value{
updated instruction list
}
\description{
Helper function that is same as transiton,
but simplifies specifying independent transitions
from multiple states to a single state.
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
