% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transition.R
\name{add_transition}
\alias{add_transition}
\title{add transition in model}
\usage{
add_transition(
  peterlist,
  fromstate,
  tostates,
  meantime,
  name = "",
  chainlength = 1,
  chaintimescale = NA,
  forkprobability = NA,
  percapita = TRUE,
  metapopulation = "",
  environment = "",
  groupname = "",
  grouptype = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{fromstate}{basestate to transition from}

\item{tostates}{character vector of basestates to
transition to}

\item{meantime}{character/numeric value specifying
average time to go from fromstate to all tostates}

\item{name}{character value to allow scaling by
other functions
default to ""}

\item{chainlength}{total steps from fromstate to
tostates (i.e., boxcars or length of pitchfork)
Alters distribution of transition times during
stochastic implementation
default to 1 which is exponentially
distrubuted transition times}

\item{chaintimescale}{character/numeric vector
with length chainlength scales rate transitioning
between steps along chains. Allows deviating
transition time distribution from a gamma distributions
default to NA which is same transition times between
all chained states (e.g., gamma distributed wait times)}

\item{forkprobability}{character/numeric vector
specifying relative weight to transtion to
different tostates when multiple are given
default to NA which specifies equal probability
transitioning to each final state}

\item{percapita}{specifies rates are given as per
capita value so the total change to population
should be scaled by the tostate
default is TRUE}

\item{metapopulation}{character vector of metapopulation
names this transitions occurs in.
default "" is which specifies all metapopulations}

\item{environment}{character vector of environment
names this transitions occurs in.
default is "" which specifies all environments}

\item{groupname}{specifies which stratified groups
transition
default is "" which specifies all groups.}

\item{grouptype}{specifies which stratified group
types transition
default is "" which specifies all grouptypes.}
}
\value{
updated instruction list
}
\description{
These assume a "pitchfork" shape to transtions
E.g., recovery/vaccination/hospitalization
to 1 or multiple states
}
