% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_travel.R
\name{add_travel}
\alias{add_travel}
\title{add travel between metapopulations in model}
\usage{
add_travel(
  peterlist,
  rates,
  metapop_from = "",
  metapop_to = "",
  travelstates = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{rates}{character/numeric vector movement rate/s}

\item{metapop_from}{character vector of origin
metapopulations
default is "" which specifies all metapopulations}

\item{metapop_to}{character vector of destination
metapopulations. default is "" which specifies all
metapopulations}

\item{travelstates}{character vector of basestates that
trave between metapopulations
default is "" which specifies all basestates}
}
\value{
updated instruction list
}
\description{
add travel between metapopulations in model
}
