% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpopsize_byname.R
\name{addpopsize_byname}
\alias{addpopsize_byname}
\title{Add popsize to named states
Add specified value to popsize of specified named state. Input tblpopsize
must have a numeric column called "popsize"}
\usage{
addpopsize_byname(tblpopsize, namevec)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{namevec}{named list/vector vector specifying updatedstate popsize
values to set. Does not impact unnamed states.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Add popsize to named states
Add specified value to popsize of specified named state. Input tblpopsize
must have a numeric column called "popsize"
}
