% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_growth_rate_calc.r
\name{apply_growth_rate_calc}
\alias{apply_growth_rate_calc}
\title{Calculates initial growth rate of outbreak}
\usage{
apply_growth_rate_calc(
  output,
  time_var = NULL,
  i_var = "i",
  beta = NULL,
  gamma = NULL,
  lower_time = NULL,
  upper_time = NULL
)
}
\arguments{
\item{output}{data frame output from a compartmental model}

\item{time_var}{character string specifying the name of the time variable,
defaults to "time"}

\item{i_var}{character string specifying the name of the infected state
variable of interest, defaults to "i"}

\item{beta}{numeric value for transmission rate, default is NULL}

\item{gamma}{numeric value for recovery rate, default is NULL}

\item{lower_time}{numeric value for time step on which will serve as lower
bound in initial growth rate calculation, default is NULL}

\item{upper_time}{numeric value for time step on which will serve as upper
bound in initial growth rate calculation, default is NULL}
}
\value{
prints initial growth rate(s) to console (numeric) and returns one
or more plots of the simulated observations, calculated theoretical growth
rate in the user-specified time range, and observed growth rate based on the
calculated growth rate r.
}
\description{
Helper function used by \code{\link[=calculate_initial_growth_rate]{calculate_initial_growth_rate()}} to
calculate the initial growth rate for the outbreak from
user supplied compartmental model output by regressing the log of the
incidence on time to estimate the rate of exponential increase r.
This r can then be used to estimate R0 with the relationship R0 = V*r + 1
where V is the serial interval.
}
\details{
Currently this calculation is for an SIR model and uses the equation
exp(i0) * exp((beta - gamma) * t)
}
\examples{
\dontrun{
# Deterministic Model
out <- run_sir(
  init = c(s = 1e05 - 1, i = 1, r = 0),
  time = seq(0.1, 100, by = 0.1),
  parms = c(beta = 0.5, gamma = 0.1)
)
calculate_initial_growth_rate(out,
  time_var = "time",
  beta = 0.5, gamma = 0.1,
  lower_time = 10,
  upper_time = 25
)

# Stochastic Model with 10 simulations
init_vals <- c(s = 499, i = 1, r = 0)
change_matrix <- matrix(
  c(
    -1, 0, # S -> I #nolint
    +1, -1, # I -> R #nolint
    0, +1
  ),
  nrow = length(init_vals),
  byrow = TRUE,
  dimnames = list(names(init_vals), NULL)
)

modelout <- run_gen_stochastic(
  parms_vec = c(
    beta = 0.5, gamma = 0.1
  ),
  propensity_fns = c(
    "(beta * s * i)/n",
    "gamma * i"
  ),
  init_vals = init_vals,
  n_timesteps = 100,
  change_matrix = change_matrix,
  n_sims = 10
)
plot_stoch_model(modelout, compartments = c("s", "i", "r"))
calculate_initial_growth_rate(modelout,
  time_var = "t", i_var = "i",
  beta = 0.5, gamma = 0.1,
  lower_time = 0,
  upper_time = 25
)
}
}
