% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_modifiers.r
\name{apply_modifiers}
\alias{apply_modifiers}
\title{Helper function to apply intervention modifiers to an expanded transition
matrix}
\usage{
apply_modifiers(expanded_trans_matrix, modifier, time, subgroup_combinations)
}
\arguments{
\item{expanded_trans_matrix}{an expanded matrix of transition rates between
for each state (compartment) and subgroup combination}

\item{modifier}{user supplied function or matrix to apply modifications to
transition matrix}

\item{time}{vector of time steps to run the model over}

\item{subgroup_combinations}{a data frame containing all combinations of
subgroups for stratification}
}
\value{
a matrix of modified transition rates between states (compartments)
}
\description{
Creates a modifier matrix of values of the same dimension as the current
transition matrix based on the user supplied modifier function, and
multiplies the modifier and transition matrix to produce a matrix of modified
transition rates between states (compartments).
}
