% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox1.R
\name{bbox1}
\alias{bbox1}
\title{A black-box computer model}
\usage{
bbox1(x1, x2)
}
\arguments{
\item{x1}{A scalar value between -1.5 and 2.5, inclusive}

\item{x2}{A scalar value between -3 and 3, inclusive}
}
\value{
The evaluation of running the black-box computer model at input (x1, x2).
\itemize{
  \item obj:  A scalar objective function value
  \item con:  A vector of constraint function values
}
}
\description{
A black-box computer model that evaluates the objective and constraint functions for a given input value.
}
\note{
A solution is feasible only if of all of the constraint functions values are less than or equal to 0.
}
\examples{
### Running the black-box computer model at x1 = 1, x2 = 2
### obj = -1.5 and con = (-3.5,-0.25)
### (x1,x2) = (1,2) is a feasible solution

bbox1(1,2)
}
