% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox5.R
\name{bbox5}
\alias{bbox5}
\title{A black-box computer model}
\usage{
bbox5(x1, x2, x3)
}
\arguments{
\item{x1}{A scalar value between -pi and pi, inclusive}

\item{x2}{A scalar value between -pi and pi, inclusive}

\item{x3}{A scalar value between -pi and pi, inclusive}
}
\value{
The evaluation of running the black-box computer model at input (x1, x2, x3).
\itemize{
  \item obj:  A scalar objective function value
}
}
\description{
A black-box computer model that evaluates the objective function for a given input value.
}
\examples{
### Running the black-box computer model at x1 = 1, x2 = 1, x3 =1
### obj = 5.840058

bbox5(1,1,1)
}
