% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox6.R
\name{bbox6}
\alias{bbox6}
\title{A black-box computer model}
\usage{
bbox6(x1)
}
\arguments{
\item{x1}{A scalar value between 0 and 10, inclusive}
}
\value{
The evaluation of running the black-box computer model at input x1.
\itemize{
  \item obj:  A scalar objective function value
  \item con:  A vector of constraint function values
}
}
\description{
A black-box computer model that evaluates the objective and constraint functions for a given input value.
}
\note{
A solution is feasible only if the all of the constraint function values are less than or equal to 0.
}
\examples{
### Running the black-box computer model at x1 = 1
### obj = 0.926574 and con = (0.4259819,0.6472136)
### x1 = 1 is not a feasible solution

bbox6(1)
}
