% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox7.R
\name{bbox7}
\alias{bbox7}
\title{A black-box computer model}
\usage{
bbox7(x1, x2, x3, x4, x5, x6, x7, x8)
}
\arguments{
\item{x1}{A scalar value between 0 and 1, inclusive}

\item{x2}{A scalar value between 0 and 1, inclusive}

\item{x3}{A scalar value between 0 and 1, inclusive}

\item{x4}{A scalar value between 0 and 1, inclusive}

\item{x5}{A scalar value between 0 and 1, inclusive}

\item{x6}{A scalar value between 0 and 1, inclusive}

\item{x7}{A scalar value between 0 and 1, inclusive}

\item{x8}{A scalar value between 0 and 1, inclusive}
}
\value{
The evaluation of running the black-box computer model at input (x1, x2, x3, x4, x5, x6, x7, x8).
\itemize{
  \item obj:  A scalar objective function value
  \item con:  A vector of constraint function values
}
}
\description{
A black-box computer model that evaluates the objective and constraint functions for a given input value.
}
\note{
A solution is feasible only if all of the constraint function values are less than or equal to 0.
}
\examples{
### Running the black-box computer model at x1 = 0, x2 = 0, x3 = 0, x4 = 0,
###                                         x5 = 0, x6 = 0, x7 = 0, x8 = 0
### obj = 1 and con = (0.2,-0.8)
### (x1,x2,x3,x4,x5,x6,x7,x8) = (0,0,0,0,0,0,0,0) is not a feasible solution

bbox7(0,0,0,0,0,0,0,0)
}
