% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate_change_rates.R
\name{calculate_change_rates}
\alias{calculate_change_rates}
\title{calculate_change_rates}
\usage{
calculate_change_rates(state, comp_names, trans_matrix)
}
\arguments{
\item{state}{holds the current state of the compartments}

\item{comp_names}{names of the compartments}

\item{trans_matrix}{matrix of transition rates between each of the
compartments}
}
\value{
calculated change rates
}
\description{
Helper function to calculate change rates for in and out flows in the
compartmental model
}
\examples{
\dontrun{
change_rates <- calculate_change_rates(
  state, comp_names,
  current_trans_matrix
)
}
}
