% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cuminf.R
\name{calculate_cuminf}
\alias{calculate_cuminf}
\title{Calculates cumulative incidence on state variable}
\usage{
calculate_cuminf(output, i_var = "i")
}
\arguments{
\item{output}{data frame output from a compartmental model}

\item{i_var}{character string specifying the name of the state
variable of interest, defaults to "i"}
}
\value{
data frame with additional column cum_inf that gives the cumulative
incidence at that time (numeric)
}
\description{
Helper function to calculate the cumulative sum of a state value,
specifically intended for summing the cumulative number of infections at a
time step and giving the cumulative sum (cumulative incidence) in a new
data frame column called cum_inf.
}
