% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_adaptivetau_args.r
\name{check_adaptivetau_args}
\alias{check_adaptivetau_args}
\title{Validate adaptivetau argument values}
\usage{
check_adaptivetau_args(transitions)
}
\arguments{
\item{transitions}{matrix object with transition information between states}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate adaptivetau argument values, used as part of a
validation function
}
