% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gillespie_args.r
\name{check_gillespie_args}
\alias{check_gillespie_args}
\title{Validate Gillespie argument values}
\usage{
check_gillespie_args(propensity_fns, change_matrix)
}
\arguments{
\item{propensity_fns}{the transition rate equations}

\item{change_matrix}{matrix that gives transition information between states}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate Gillespie argument values, used as part of a validation
function
}
