% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_init_vals.R
\name{check_init_vals}
\alias{check_init_vals}
\title{Validate initial values}
\usage{
check_init_vals(init_vals)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate initial values, used as part of a validation
}
