% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_initvals_compnames.R
\name{check_initvals_compnames}
\alias{check_initvals_compnames}
\title{Validate number of initial values matches number of compartment names}
\usage{
check_initvals_compnames(init_vals, comp_names)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}

\item{comp_names}{names of the compartments}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate that the number of initial values matches number
of compartment names, used as part of a validation
function
}
