% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_int_start_end.r
\name{check_intervention_times}
\alias{check_intervention_times}
\title{Validate intervention start and end times}
\usage{
check_intervention_times(intervention_start_time, intervention_end_time)
}
\arguments{
\item{intervention_start_time}{time at which intervention will start}

\item{intervention_end_time}{time at which intervention will end}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate intervention start and end times, used as part
of a validation function
}
