% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_modifier1.r
\name{check_modifier1}
\alias{check_modifier1}
\title{Validate modifier 1}
\usage{
check_modifier1(modifier)
}
\arguments{
\item{modifier}{single value, matrix, or function to be applied to transition
matrix during an intervention}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate modifier values, used as part
of a validation function. A modifier should be one of: a single numeric
value, a matrix of numeric values, or a function.
}
