% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nsims.r
\name{check_nsims}
\alias{check_nsims}
\title{Validate number of simulations step value}
\usage{
check_nsims(n_sims)
}
\arguments{
\item{n_sims}{number of simulations desired}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate number of simulations step value, used as part
of a validation function
}
