% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parms_vec.R
\name{check_parms_vec}
\alias{check_parms_vec}
\title{Validate parameter values}
\usage{
check_parms_vec(parms_vec)
}
\arguments{
\item{parms_vec}{vector of parameter values}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate time step value, used as part of a validation
function
}
