% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_subgroups_list.r
\name{check_subgroups_list}
\alias{check_subgroups_list}
\title{Validate subgroups list}
\usage{
check_subgroups_list(subgroups_list)
}
\arguments{
\item{subgroups_list}{list of one or more vectors of subgroup names to be
used for stratification}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate subgroups, used as part of a validation
function
}
