% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_time_step.R
\name{check_time_step}
\alias{check_time_step}
\title{Validate time step value}
\usage{
check_time_step(n_timesteps)
}
\arguments{
\item{n_timesteps}{number of time steps}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate time step value, used as part of a validation
function
}
