% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_times.r
\name{check_times}
\alias{check_times}
\title{Validate time values}
\usage{
check_times(times)
}
\arguments{
\item{times}{vector of time steps to run the model over}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate time values, used as part of a validation
function
}
