% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trans_mtx.r
\name{check_trans_mtx}
\alias{check_trans_mtx}
\title{Validate transition matrix}
\usage{
check_trans_mtx(trans_matrix, comp_names)
}
\arguments{
\item{trans_matrix}{matrix of transition rates between each of the
compartments}

\item{comp_names}{names of the compartments}
}
\value{
stops with information or indicates checks passed
}
\description{
Helper function to validate transition matrix, used as part of a validation
function
}
