% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceinput2matchshape.R
\name{coerceinput2matchshape}
\alias{coerceinput2matchshape}
\title{Converts input list to nested lists of lists}
\usage{
coerceinput2matchshape(queryinput, listofvectors)
}
\arguments{
\item{queryinput}{character vector to be replicated}

\item{listofvectors}{list of lists whose shape is
use to coerce character vector}
}
\value{
list of lists with inputbasestate as values
}
\description{
Useful when assigning multiple environments
and metapopulations
}
