% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceinput2matchshape_basestate.R
\name{coerceinput2matchshape_basestate}
\alias{coerceinput2matchshape_basestate}
\title{Converts given base state list to nested lists of lists.}
\usage{
coerceinput2matchshape_basestate(inputbasestate, listofvectors)
}
\arguments{
\item{inputbasestate}{character vector to be replicated}

\item{listofvectors}{list of lists whose shape is
use to coerce character vector}
}
\value{
list of lists with inputbasestate as values
}
\description{
Useful when assigning multiple environments and metapopulations
}
