% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc2inc.R
\name{cuminc2inc}
\alias{cuminc2inc}
\title{Convert cumulative incidence to incidence}
\usage{
cuminc2inc(t, cuminc, keepfirst = TRUE)
}
\arguments{
\item{t}{compiled model output}

\item{cuminc}{named list/vector vector specifying updatedstate population
sizes}

\item{keepfirst}{logical whether first element of cuminc should be included
as incident}
}
\value{
numeric vector of incidence if keepfirst and
a named list with time and incidence vectors
}
\description{
Requires times vector and cumulative incidence vector, finds where cumulative
incidence changes, and then outputs when the change occurs and the magnitude
}
