% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_currentstate.R
\name{define_currentstate}
\alias{define_currentstate}
\title{Set population states for compiled
model using all states}
\usage{
define_currentstate(outlist, namedvector = c())
}
\arguments{
\item{outlist}{compiled model output}

\item{namedvector}{numeric vector specifying initial populations
for named basestates
default is c() which specifies all 0 population
for each basestate}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (X0)
}
\description{
Sets conditions globally based on current states.
}
