% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_initialstate.R
\name{define_initialstate}
\alias{define_initialstate}
\title{Set initial population states for compiled
model using base states}
\usage{
define_initialstate(compiledmodel, basestatevector = c())
}
\arguments{
\item{compiledmodel}{compiled model output}

\item{basestatevector}{numeric vector specifying initial populations
for named basestates
default is c() which specifies all 0 population
for each basestate}
}
\value{
tibble with columns of updated state names
(updatedstates) and initial conditions (X0)
}
\description{
Sets initial conditions globally based on base states.
}
\seealso{
Other initial_conditions:
\code{\link{define_popsize}()},
\code{\link{define_states}()},
\code{\link{odin_initial_states}()},
\code{\link{output_initialstate}()}
}
\concept{initial_conditions}
