% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_metapopulations.R
\name{define_metapopulations}
\alias{define_metapopulations}
\title{Define metapopulation names}
\usage{
define_metapopulations(
  peterlist,
  metapop_names = "",
  environment_names = "",
  interactionscale = 1,
  transitionscale = 1,
  basestates = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{metapop_names}{character vector specifying
metapopulation names
default is ""}

\item{environment_names}{character vector specifying
environment names
default is ""}

\item{interactionscale}{numeric vector that scales
interactions in
each specified metapopulation
default is 1}

\item{transitionscale}{numeric vector that scales
transitions in
each specified metapopulation
default is 1}

\item{basestates}{character vector specifying
which basestates exist
in each metapopulations
default is "" which specifies all basestates}
}
\value{
model instructions
}
\description{
Removes all prior add_metapopulation instructions and sets
them according to input
}
