% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_popsize.R
\name{define_popsize}
\alias{define_popsize}
\title{Set population states for compiled
model using all states}
\usage{
define_popsize(compiledmodel, inputpops = c())
}
\arguments{
\item{compiledmodel}{compiled model output}

\item{inputpops}{named list/vector vector specifying updatedstate population
sizes
Unspecified populations are set to 0}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Sets conditions globally based on current states.
}
