% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_states.R
\name{define_states}
\alias{define_states}
\title{Initializes model instructions with given base states}
\usage{
define_states(base_state, metapopulation = "")
}
\arguments{
\item{base_state}{character vector of base states.}

\item{metapopulation}{character vector of metapopulation names
default is ""}
}
\value{
list of model instructions
}
\description{
Initializes model instructions with given base states
}
\seealso{
Other initial_conditions:
\code{\link{define_initialstate}()},
\code{\link{define_popsize}()},
\code{\link{odin_initial_states}()},
\code{\link{output_initialstate}()}
}
\concept{initial_conditions}
