% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaytime_byfunction.R
\name{delaytime_byfunction}
\alias{delaytime_byfunction}
\title{Delay times according to function}
\usage{
delaytime_byfunction(t, delayfunction, sortlogic = TRUE)
}
\arguments{
\item{t}{vector of numerical times}

\item{delayfunction}{non-negative numeric or function that outputs
non-negative value to add to time
If a function, it should not have any input arguments}

\item{sortlogic}{optional logic whether to sort the new times. Sorting is
useful for outputting time series. Not sorting is useful to keep the order
according to the latent counts associated with the times
default is TRUE which designates sorting}
}
\value{
updated times
}
\description{
Function can be deterministic (e.g., a single number) or stochastic (e.g.,
sampling from a distribution like rgamma()).
}
