% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaytime_exp.R
\name{delaytime_exp}
\alias{delaytime_exp}
\title{Randomly delay times according to exponential distribution}
\usage{
delaytime_exp(t, meandelay, sortlogic = TRUE)
}
\arguments{
\item{t}{vector of numerical times}

\item{meandelay}{average time to add, sets exponential distribution to sample
from}

\item{sortlogic}{optional logic whether to sort the new times. Sorting is
useful for outputting time series. Not sorting is useful to keep the order
according to the latent counts associated with the times
default is TRUE which designates sorting}
}
\value{
updated times
}
\description{
Randomly delay times according to exponential distribution
}
