% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaytime_gamma.R
\name{delaytime_gamma}
\alias{delaytime_gamma}
\title{Randomly delay times according to Gamma distribution}
\usage{
delaytime_gamma(t, meandelay, shape, sortlogic = TRUE)
}
\arguments{
\item{t}{vector of numerical times}

\item{meandelay}{average time to add, sets mean of erlang distribution to
sample from}

\item{shape}{number of steps in for the analogous chained transition}

\item{sortlogic}{optional logic whether to sort the new times. Sorting is
useful for outputting time series. Not sorting is useful to keep the order
according to the latent counts associated with the times
default is TRUE which designates sorting}
}
\value{
updated times
}
\description{
Note: if an integer shape is input then the sampled gamma distribution
matches an erlang distribution, which is the distribution of times resulting
from chained transition processes.
}
