% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_initial_vals.r
\name{expand_init_vals}
\alias{expand_init_vals}
\title{Helper function to expand initial values based on subgroup combinations}
\usage{
expand_init_vals(init_vals_base, num_combinations)
}
\arguments{
\item{init_vals_base}{the initial values supplied by the user for each state
(compartment) before stratification}

\item{num_combinations}{the total number of all subgroup combinations}
}
\value{
an expanded vector of initial values for each state (compartment) and
subgroup combination
}
\description{
Takes in a vector of initial values for the user defined names and number of
model states (compartments) and determines all the combinations of states and
user defined subgroups and distributes individuals evenly across all
combinations.
}
