% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_trans_mtx.r
\name{expand_trans_mtx}
\alias{expand_trans_mtx}
\title{Helper function to expand transition matrix based on number of subgroup
combinations}
\usage{
expand_trans_mtx(trans_matrix_base, num_combinations)
}
\arguments{
\item{trans_matrix_base}{the initial matrix providing transition rates
between states (compartments) supplied by the user before stratification}

\item{num_combinations}{the total number of all subgroup combinations}
}
\value{
an expanded matrix of transition rates between for each state
(compartment) and subgroup combination.
}
\description{
Takes in the user supplied transition matrix of rates of flow between model
states (compartments) and expands the transition matrix to dimensions of all
combinations of states and subgroups
}
