% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixinput_chains.R
\name{fixinput_chains}
\alias{fixinput_chains}
\title{Convert chains input to named list}
\usage{
fixinput_chains(chains, tblwgroupcols)
}
\arguments{
\item{chains}{vector or named list of chains values are elements of
input table}

\item{tblwgroupcols}{table that has chainid columns with chains as
values}
}
\value{
chains coerced into named list with respective chainid as name
}
\description{
chains input can be vector or named list, depending on model complexity.
This function coerces input so that it is a named list with names matching
table columns that contain input values. The resulting consistent format
simplifies manipulating user input across functions.
}
