% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixinput_groupnames.R
\name{fixinput_groupnames}
\alias{fixinput_groupnames}
\title{Convert groupnames input to named list}
\usage{
fixinput_groupnames(groupnames, tblwgroupcols)
}
\arguments{
\item{groupnames}{vector or named list of groupnames. values are elements of
input table}

\item{tblwgroupcols}{table that has grouptype columns with groupnames as
values}
}
\value{
groupnames coerced into named list with respective grouptype as name
}
\description{
groupnames input can be vector or named list, depending on model complexity.
This function coerces input so that it is a named list with names matching
table columns that contain input values. The resulting consistent format
simplifies manipulating user input across functions.
}
