% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rates.r
\name{gen_rates}
\alias{gen_rates}
\title{Calculates modified rates of change for in/out flows from compartments}
\usage{
gen_rates(time, state, parms)
}
\arguments{
\item{time}{vector of time steps to run the model over}

\item{state}{holds the current state of the compartments}

\item{parms}{additional parameters needed such as comp_names (compartment
names), trans_matrix (transition rate matrix), intervention_start_time,
intervention_end_time, modifier_matrix (values to be multiplied against
transition matrix)}
}
\value{
list object with columns for time and each compartment
}
\description{
Helper function to calculate modified change rates for in and out flows in
a generalized compartmental model with intervention
\code{\link[=run_gen_deterministic]{run_gen_deterministic()}}.
Helper function \code{\link[=apply_modifiers]{apply_modifiers()}} is used during interventions
}
