% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comp_names.r
\name{generate_comp_names}
\alias{generate_comp_names}
\title{Helper function to generate compartment names for all subgroup combinations}
\usage{
generate_comp_names(comp_names_base, subgroup_combinations)
}
\arguments{
\item{comp_names_base}{the initial values of compartment names supplied by
the user for each state (compartment) before stratification}

\item{subgroup_combinations}{a data frame containing all combinations of
subgroups for stratification}
}
\value{
an expanded vector of compartment names for each state (compartment)
and subgroup combination.
}
\description{
Takes in a vector of initial values for the user defined names of model
states (compartments) and determines all the name combinations of states and
user defined subgroups
}
