% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie.R
\name{gillespie}
\alias{gillespie}
\title{Generate model simulation using gillespie algorithm}
\usage{
gillespie(listmodel, params, x0, tvec = NA, maxtransitions = NA)
}
\arguments{
\item{listmodel}{named list generated by compilemodel}

\item{params}{named list of parameter values}

\item{x0}{named list specifying initial (integer)
populations}

\item{tvec}{optional simulation time.
default is NA which specfies running until
process rates are zero}

\item{maxtransitions}{optionally limit total
number of population changes
default is NA which specfies running
until process rates are zero}
}
\value{
tibble of dynamics with column t and
columns of state names
}
\description{
Work in progress. fix list2env.
}
