% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabfromtbl.R
\name{grabfromtbl}
\alias{grabfromtbl}
\title{Grabs rows that matches values in (named)list}
\usage{
grabfromtbl(namedlist, tbl2grab)
}
\arguments{
\item{namedlist}{(named)list of groups to grab. If names are given they
correspond to types}

\item{tbl2grab}{table with that has columns with values matching those in the
(named)list. If a named list is input, then some column names match the
namedlist names.}
}
\value{
table with rows that contain the named values
}
\description{
Names of input named list are column names.
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
