% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcombinations.R
\name{groupcombinations}
\alias{groupcombinations}
\title{Generates all combinations groups with different grouptypes}
\usage{
groupcombinations(combinethesetypes, tblgroup)
}
\arguments{
\item{combinethesetypes}{list of character vectors that
specify grouptypes to combine groups.}

\item{tblgroup}{tibble specifying groups in each group type}
}
\value{
tibble where each row is a combination of groups
with grouptype columns and scaled process rates
}
\description{
Combines groups based on input group types. Useful when compiling to build
updated state table.
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
