% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcombinationsacrosstypes.R
\name{groupcombinationsacrosstypes}
\alias{groupcombinationsacrosstypes}
\title{Generates all combinations groups with different grouptypes}
\usage{
groupcombinationsacrosstypes(combinethesetypes, tblgroup)
}
\arguments{
\item{combinethesetypes}{list of character vectors that
specify grouptypes to combine groups.}

\item{tblgroup}{tibble specifying groups in each group type}
}
\value{
tibble where each row is a combination of groups
with grouptype columns and scaled process rates
}
\description{
Useful when stratifying populations in multiple ways.
}
