% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intlogic.R
\name{intlogic}
\alias{intlogic}
\title{Determine if popsize has integer values}
\usage{
intlogic(tblpopsize)
}
\arguments{
\item{tblpopsize}{table with popsize column}
}
\value{
logical for containing integers
}
\description{
Determine if popsize has integer values
}
