% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_intervention_period.R
\name{is_intervention_period}
\alias{is_intervention_period}
\title{Check time within intervention period}
\usage{
is_intervention_period(time, start_time, end_time)
}
\arguments{
\item{time}{vector (type: double) containing time steps to run the model over}

\item{start_time}{time step (type: double) at which intervention will start}

\item{end_time}{time step (type: double) at which intervention will end}
}
\value{
logical response if the time supplied falls in the bounds of an
intervention time
}
\description{
Helper function to determine if the time supplied falls inside or on the
supplied bounds (inclusive)
}
\examples{
\dontrun{
yn_intervention <- is_intervention_period(
  time, intervention_start_time,
  intervention_end_time
)
}
}
