% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintainpopsizetype.R
\name{maintainpopsizetype}
\alias{maintainpopsizetype}
\title{Converts column to int if values have no remainder and total is larger than 1}
\usage{
maintainpopsizetype(tbl2maintain, col2maintain)
}
\arguments{
\item{tbl2maintain}{table with column of values}

\item{col2maintain}{named column to convert values to int
Unspecified populations are set to 0}
}
\value{
input tbl with column converted to int if values have no remainder
}
\description{
Useful for determining how to sample populations
}
