% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metapopulations.R
\name{make_metapopulations}
\alias{make_metapopulations}
\title{Set metapopulations in model}
\usage{
make_metapopulations(
  metapopulation = "",
  scaleinteractions = 1,
  scaletransitions = 1,
  scaleprocessbyname = list(),
  scaleprocessbygroup = list(),
  basestates = "",
  groups = ""
)
}
\arguments{
\item{metapopulation}{vector of metapopulation names
default is ""}

\item{scaleinteractions}{vector same length as
input metapopulations/environment_names that
scales all interactions in given metapop/environment
default is 1}

\item{scaletransitions}{vector same length as
input metapopulations/environment_names that
scales all transitions in given metapop/environment
default is 1}

\item{scaleprocessbyname}{named list with either vectors of the same length
as the input metapopulations or length 1. Scales named processes in
each metapopulation.}

\item{scaleprocessbygroup}{named list with either vectors of the same length
as the input metapopulations or length 1. Scales grouped named processes in
each metapopulation.}

\item{basestates}{specifies which states can
visit metapopulation/environment.
default is "" which specifies all basestates}

\item{groups}{specifies which groups can visit metapopulation. This can be a
named list with grouptypes as names and groups as elements
or a vector of group names
default is "" which species all groups}
}
\value{
tibble combining all information.
}
\description{
Checks user input and outputs tibble to add metapopulations to instructions
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
