% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenewvariablenamesfromrow.R
\name{makenewvariablenamesfromrow}
\alias{makenewvariablenamesfromrow}
\title{Generates new state names from basestates}
\usage{
makenewvariablenamesfromrow(tibblerow)
}
\arguments{
\item{tibblerow}{tibble row to generate state names from
Ignore columns with all NA.}
}
\value{
string of updated name
}
\description{
Pastes basestate names with column names and column
values, delimited by _
}
