% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchandpaste.R
\name{matchandpaste}
\alias{matchandpaste}
\title{Convert vector to input length}
\usage{
matchandpaste(inputvector, desiredlength)
}
\arguments{
\item{inputvector}{character or numeric vector.}

\item{desiredlength}{output vector length}
}
\value{
lengthened vector same type as input
}
\description{
Replicates length one vector and ensures n-length
vectors are same size as given length.
Same as matchlength.R
}
