% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchandpaste_basestate.R
\name{matchandpaste_basestate}
\alias{matchandpaste_basestate}
\title{Convert list/vector to input length, keeps values as lists if}
\usage{
matchandpaste_basestate(inputvector, desiredlength)
}
\arguments{
\item{inputvector}{character/numeric vector or list}

\item{desiredlength}{output vector length}
}
\value{
lengthened vector/list same type as input
}
\description{
Replicates length one vector and ensures n-length vectors
are same size as given length. respects if input
values are lists.
}
