% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchlength.R
\name{matchlength}
\alias{matchlength}
\title{Convert vector to input length}
\usage{
matchlength(vec, desiredlength)
}
\arguments{
\item{vec}{character or numeric vector.}

\item{desiredlength}{output vector length}
}
\value{
lengthened vector same type as input
}
\description{
Replicates length one vector and ensures n-length
vectors are same size as given length.
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
