% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2desolvefunction.R
\name{model2desolvefunction}
\alias{model2desolvefunction}
\title{Generate deSolve input function from compiled model}
\usage{
model2desolvefunction(outlist)
}
\arguments{
\item{outlist}{compiled model including
name "modeloutsructions"}
}
\value{
deSolve input function.
}
\description{
Generate deSolve input function from compiled model
}
