% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_states.R
\name{modify_states}
\alias{modify_states}
\title{Modify state values}
\usage{
modify_states(
  vec,
  modifications = list(),
  adjust_names = NULL,
  adjust_value = 0
)
}
\arguments{
\item{vec}{a named numeric vector containing the current values of states
to be modified}

\item{modifications}{(optional) a user supplied list of named state
compartments and corressponding new values for the named compartments,
default is empty list, list()}

\item{adjust_names}{a character string or vector of character strings of
the names or prefixes of the state compartments for which the user would
like to adjust the corresponding values}

\item{adjust_value}{a numeric value or vector of numeric values of
the revised values of the state compartments for which the user would like to
adjust the corresponding values}
}
\value{
named vector of compartment state values
}
\description{
Enables user to make fine grained updates to compartment state values.
}
