% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_modify_trans_mtx.R
\name{modify_trans_mtx}
\alias{modify_trans_mtx}
\title{modify_trans_mtx}
\usage{
modify_trans_mtx(trans_matrix, intervention_modifiers)
}
\arguments{
\item{trans_matrix}{matrix of transition rates between each of the
compartments}

\item{intervention_modifiers}{value to be multiplied against specific
values in transition matrix}
}
\value{
modified transmission matrix
}
\description{
Helper function to modify the transition matrix by applying a modifier to
specific user defined compartments
}
\examples{
\dontrun{
current_trans_matrix <- modify_trans_mtx(
  trans_matrix,
  intervention_modifiers
)
}
}
