% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movefracpopsize_byfeature.R
\name{movefracpopsize_byfeature}
\alias{movefracpopsize_byfeature}
\title{Move fraction of popsize between states matching input features}
\usage{
movefracpopsize_byfeature(
  tblpopsize,
  fracpopsize,
  basestates_before = NA,
  basestates_after = NA,
  basestates = NA,
  groupnames_before = NA,
  groupnames_after = NA,
  groupnames = NA,
  metapopulation_before = NA,
  metapopulation_after = NA,
  metapopulation = NA,
  chains_before = NA,
  chains_after = NA,
  chains = NA,
  samplerandomly = NA
)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{fracpopsize}{fraction of population to move from each popsize that
matches input features}

\item{basestates_before}{character vector of basestate to move movepopsize
from. Many values may be provided.
default is NA which ignores this input}

\item{basestates_after}{character vector of basestate to move movepopsize
into. At most one value may be provided.
default is NA which ignores this input}

\item{basestates}{vector of basestates specifying moving populations.
default is NA which designates all basestates, unless user provides _before
and _after input arguments}

\item{groupnames_before}{character vector/list of groupnames to move
movepopsize from. Many values may be specified.
default is NA which ignores this input}

\item{groupnames_after}{character vector/list of groupnames to move
movepopsize into. Single value must be specified (altogether,
or for each name in list)
default is NA which ignores this input}

\item{groupnames}{list of groupnames specifying moving populations.
default is NA which designates all basestates, unless user provides _before
and _after input arguments}

\item{metapopulation_before}{character vector of metapopulation to move
movepopsize from. Many values may be provided.
default is NA which ignores this input}

\item{metapopulation_after}{character vector of metapopulations to move
movepopsize into. At most one value may be provided.
default is NA which ignores this input}

\item{metapopulation}{vector of metapopulations specifying moving
populations. default is NA which designates all metapopulation, unless user
provides _before and _after input arguments}

\item{chains_before}{character vector/list of chains steps to move
movepopsize from. Many values may be specified.
default is NA which ignores this input}

\item{chains_after}{character vector/list of chains steps to move movepopsize
into. Single value must be specified (altogether, or for each name in list)
default is NA which ignores this input}

\item{chains}{list of chains step specifying moving populations.
default is NA which designates all basestates, unless user provides _before
and _after input arguments}

\item{samplerandomly}{logical specifying whether population moved is
determined by binomial sampling (TRUE and integer populations) or by the
fractional amount (FALSE) with rounding when integer populations.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Multi-argument function that flexibly moves specified fractions of
populations between states. Exactly how much population moves depends on
whether the fraction to move is sampled deterministically or stochastically
via the samplerandomly input argument. Users can specify
features that change before and after the move, along with features that do
not change before and after the move. Multiple states can move into a single
state, allowing a more-to-fewer mapping across the whole popsize table.
Checks ensure proper input argument combinations.
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
