% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movefracpopsize_byname.R
\name{movefracpopsize_byname}
\alias{movefracpopsize_byname}
\title{Move fraction of popsize for named states}
\usage{
movefracpopsize_byname(
  tblpopsize,
  fracpopsize,
  namebefore,
  nameafter,
  samplerandomly
)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize column}

\item{fracpopsize}{fraction of population to move between states}

\item{namebefore}{updatedstate value(s) in tblupdatedstate to remove popsize}

\item{nameafter}{updatedstate value in tblupdatedstate to add popsize}

\item{samplerandomly}{logical specifying whether population moved is
determined by binomial sampling (TRUE and integer populations) or by the
fractional amount (FALSE) with rounding when integer populations.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Move fraction of populations between some number of named updatedstates to
another single named state.
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
