% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepopsize_byname.R
\name{movepopsize_byname}
\alias{movepopsize_byname}
\title{Move popsize for named states}
\usage{
movepopsize_byname(tblpopsize, popsizeval, namebefore, nameafter)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize column}

\item{popsizeval}{value to move between states}

\item{namebefore}{updatedstate value(s) in tblupdatedstate to remove popsize}

\item{nameafter}{updatedstate value in tblupdatedstate to add popsize}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Move populations between some number of named updatedstates and to another
single named state.
}
