% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movethresholdpopsize_byfeature.R
\name{movethresholdpopsize_byfeature}
\alias{movethresholdpopsize_byfeature}
\title{Move thresholded popsize for between states matching input features}
\usage{
movethresholdpopsize_byfeature(
  tblpopsize,
  movepopsize,
  basestates_before = NA,
  basestates_after = NA,
  basestates = NA,
  groupnames_before = NA,
  groupnames_after = NA,
  groupnames = NA,
  metapopulation_before = NA,
  metapopulation_after = NA,
  metapopulation = NA,
  chains_before = NA,
  chains_after = NA,
  chains = NA,
  movefromallchains = FALSE
)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{movepopsize}{value to make all popsizes that match input features}

\item{basestates_before}{character vector of basestate to move movepopsize
from. Many values may be provided.
default is NA which ignores this input}

\item{basestates_after}{character vector of basestate to move movepopsize
into. At most one value may be provided.
default is NA which ignores this input}

\item{basestates}{vector of basestates specifying moving populations.
default is NA which designates all basestates, unless user provides _before
and _after input arguments}

\item{groupnames_before}{character vector/list of groupnames to move
movepopsize from. Many values may be specified.
default is NA which ignores this input}

\item{groupnames_after}{character vector/list of groupnames to move
movepopsize into. Single value must be specified (altogether,
or for each name in list)
default is NA which ignores this input}

\item{groupnames}{list of groupnames specifying moving populations.
default is NA which designates all basestates, unless user provides _before
and _after input arguments}

\item{metapopulation_before}{character vector of metapopulation to move
movepopsize from. Many values may be provided.
default is NA which ignores this input}

\item{metapopulation_after}{character vector of metapopulations to move
movepopsize into. At most one value may be provided.
default is NA which ignores this input}

\item{metapopulation}{vector of metapopulations specifying moving
populations. default is NA which designates all metapopulation, unless user
provides _before and _after input arguments}

\item{chains_before}{vector/named list of chains steps to move
movepopsize from. Many values may be specified.
default is NA which ignores this input}

\item{chains_after}{vector/named list of chains steps to move movepopsize
into. Vector input must be length 1. Named values must be length 1 for
named list input, but the list itself can have length >1. Names correspond to
chainids.
default is NA which ignores this input}

\item{chains}{vector/named list of chains steps to fix when moving
populations.
default is NA which assumes moving from first step in all chains to
first step in other chains as stratified by other input. Users can specify
chains _before and _after using the respective input arguments}

\item{movefromallchains}{flag to specify moving from all steps in before
chains. The default is set to false so movements flow from first step in
before chains.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Multi-argument function that flexibly moves populations between states. Users
can specify features that change before and after the move, along with
features that do not change before and after the move, but specify
which populations to move between. Multiple states can move into a single
state, allowing a more-to-fewer mapping across the whole popsize table.
Checks ensure proper input argument combinations. Amount moved is up to
specifed movepopsize ensuring no negative populations
}
