% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movethresholdpopsize_byname.R
\name{movethresholdpopsize_byname}
\alias{movethresholdpopsize_byname}
\title{Move thresholded popsize between named states}
\usage{
movethresholdpopsize_byname(tblpopsize, popsizeval, namebefore, nameafter)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize column}

\item{popsizeval}{value to move between states}

\item{namebefore}{updatedstate value(s) in tblupdatedstate to remove popsize}

\item{nameafter}{updatedstate value in tblupdatedstate to add popsize}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Move populations between some number of named updatedstates and to another
single named state. Amount moved is up to specifed popsizeval ensuring no
negative populations
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
