% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedlist2tibblerow.R
\name{namedlist2tibblerow}
\alias{namedlist2tibblerow}
\title{Create 1 row tibble from list where input
values are treated as lists}
\usage{
namedlist2tibblerow(namedlist)
}
\arguments{
\item{namedlist}{named list}
}
\value{
1 row tibble with names as columns
and values as lists
}
\description{
Create 1 row tibble from list where input
values are treated as lists
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
