% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize2mean1.R
\name{normalize2mean1}
\alias{normalize2mean1}
\title{Normalize vector to have mean 1}
\usage{
normalize2mean1(inputvector)
}
\arguments{
\item{inputvector}{Numeric or character vector.}
}
\value{
normalized vector of same type
}
\description{
Normalize vector to have mean 1
}
