% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odegenerator.R
\name{odegenerator}
\alias{odegenerator}
\title{Generate deSolve function}
\usage{
odegenerator(expressions)
}
\arguments{
\item{expressions}{expressions of instantaneous
population changes}
}
\value{
deSolve input function
}
\description{
Generate deSolve function
}
