% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_draws.R
\name{odin_draws}
\alias{odin_draws}
\title{Draws distributions for numbers changing between compartments/states
Refer to https://mrc-ide.github.io/odin.dust/articles/sir_models.html
for format examples}
\usage{
odin_draws(compiled_model, states_vec)
}
\arguments{
\item{compiled_model}{compiled model object
created by compModels function compilemodel()}

\item{states_vec}{vector of states in model as character strings
example: base_states <- c("S", "E", "I", "R")}
}
\value{
vector of strings formatted to describe draws of numbers
changing between states
}
\description{
In progress: need to add functionality for multinomial draws
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
