% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_initial_states.R
\name{odin_initial_states}
\alias{odin_initial_states}
\title{Create initial states for odin model}
\usage{
odin_initial_states(states_vec, initial_states_vec)
}
\arguments{
\item{states_vec}{vector of states in model as character strings
example: base_states <- c("S", "E", "I", "R")}

\item{initial_states_vec}{named vector of initial values for each state
example: initial_states_vec <- c("S" = 999, "E" = 0, "I" = 1, "R" = 0)}
}
\value{
vector of strings formatted for odin initial states
}
\description{
Format needs to be initial(state) <- value for each value
Need to provide a vector of states and a named vector of initial values.
If a state is provided with no value, it will be set to 0.
}
\seealso{
Other initial_conditions:
\code{\link{define_initialstate}()},
\code{\link{define_popsize}()},
\code{\link{define_states}()},
\code{\link{output_initialstate}()}
}
\concept{initial_conditions}
