% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_probs.R
\name{odin_probs}
\alias{odin_probs}
\title{Calculate individual transition probabilities for odin model
Refer to https://mrc-ide.github.io/odin.dust/articles/sir_models.html
for format examples}
\usage{
odin_probs(compiled_model)
}
\arguments{
\item{compiled_model}{compiled model object
created by compModels function compilemodel()}
}
\value{
vector of strings formatted to describe transition probabilities
}
\description{
Calculate individual transition probabilities for odin model
Refer to https://mrc-ide.github.io/odin.dust/articles/sir_models.html
for format examples
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
