% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_tot_pop.R
\name{odin_tot_pop}
\alias{odin_tot_pop}
\title{Create calculate total population for odin model}
\usage{
odin_tot_pop(model_pop)
}
\arguments{
\item{model_pop}{model population information from
compiled model $modelinstructions$tblntotal}
}
\value{
string formatted to calculate total population
}
\description{
Total value is called N and is the sum of provided states
Total N can then be used in model calculations
}
\seealso{
Other calculations:
\code{\link{calculate_initial_growth_rate}()},
\code{\link{report_incidence}()},
\code{\link{trackincidence_byfeature}()},
\code{\link{trackincidence_byname}()}
}
\concept{calculations}
