% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_initialstate.R
\name{output_initialstate}
\alias{output_initialstate}
\title{Converts table of initial conditions to named list}
\usage{
output_initialstate(x0tbl)
}
\arguments{
\item{x0tbl}{tibble of initial conditions
with columns
x0 and updatedstate}
}
\value{
named list with names updated state and values x0
}
\description{
Useful as input to deSolve
}
\seealso{
Other initial_conditions:
\code{\link{define_initialstate}()},
\code{\link{define_popsize}()},
\code{\link{define_states}()},
\code{\link{odin_initial_states}()}
}
\concept{initial_conditions}
