% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peterrow2dxdt.R
\name{peterrow2dxdt}
\alias{peterrow2dxdt}
\title{Get instantaneous change to population}
\usage{
peterrow2dxdt(currrow, currprocessrates)
}
\arguments{
\item{currrow}{Row of petersen matrix with integer values}

\item{currprocessrates}{character vector of process
rates to be added together}
}
\value{
String specifying total instaneous change
across the row (ie for a state)
}
\description{
For a given row of petersen matrix,
add/subtract process
rates as scaled by value in row. Useful to
generate input for deSolve
}
