% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_determin_model.R
\name{plot_determin_model}
\alias{plot_determin_model}
\title{Plots deterministic compartmental models}
\usage{
plot_determin_model(
  output,
  stratify_by = NULL,
  show_intervention_period = FALSE,
  intervention_period = NULL
)
}
\arguments{
\item{output}{data frame output from a compartmental model}

\item{stratify_by}{optional argument where user can name stratifications to
include. Each name will result in one plot per strata}

\item{show_intervention_period}{optional logical argument (default FALSE) to
add line(s) showing the start and/or end of the intervention}

\item{intervention_period}{optional argument (default NULL) to add a vector
(type: numeric) with the start and end time of an intervention}
}
\value{
ggplot2 plot of compartmental model output
}
\description{
Plots deterministic compartmental model output in using \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\examples{
\dontrun{
modelout <- run_sir(
  init = c(s = 1e05 - 1, i = 1, r = 0),
  time = seq(0.1, 100, by = 0.1),
  parms = c(beta = 0.00001, gamma = 0.1)
)
plot_determin_model(modelout)

modelout2 <- run_sir_intervention(
  init = c(s = 1e05 - 1, i = 1, r = 0),
  time = seq(0.1, 100, by = 0.1),
  parms = list(
    beta = 0.00001, gamma = 0.1,
    intervention_start_time = 10,
    intervention_end_time = 20,
    intervention_impact = 0.3
  )
)
plot_determin_model(modelout2,
  show_intervention_period = TRUE,
  intervention_period = c(10, 20)
)

init_vals_base <- c(s = 99000, e = 0, i = 10000, r = 0)
comp_names_base <- c("s", "e", "i", "r")
beta <- .9 # transmission
sigma <- .1 # progression E->I
gamma <- 0.05 # recovery
trans_matrix_base <- matrix(
  c(
    -beta, 0, 0, 0,
    beta, -sigma, 0, 0,
    0, sigma, -gamma, 0,
    0, 0, gamma, 0
  ),
  nrow = length(comp_names_base), byrow = TRUE,
  dimnames = list(comp_names_base, comp_names_base)
)
times <- seq(0, 100, by = 1)
age_grps_vector <- c("child", "adult", "elderly")
vacc_status_vector <- c("vaccinated", "unvaccinated")
gender_cat_vector <- c("male", "female")

modelout3 <- run_gen_deterministic(
  init_vals_base = init_vals_base,
  times = times,
  comp_names_base = comp_names_base,
  trans_matrix_base = trans_matrix_base,
  subgroups_list =
    list(
      age_groups = age_grps_vector,
      vaccination_statuses = vacc_status_vector,
      gender_categories = gender_cat_vector
    ),
  intervention_start_time = NULL,
  intervention_end_time = NULL,
  modifier_function = NULL
)
plot_determin_model(modelout3,
  stratify_by = c("child", "adult", "elderly")
)
}
}
