% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_model.R
\name{plot_stoch_model}
\alias{plot_stoch_model}
\title{Plots stochastic compartmental models}
\usage{
plot_stoch_model(
  output,
  compartments = c("i"),
  time_var = NULL,
  show_intervention_period = FALSE,
  intervention_period = NULL
)
}
\arguments{
\item{output}{data frame output from a compartmental model}

\item{compartments}{character string specifying which compartment to plot,
defaults to "i" for infected}

\item{time_var}{character string specifying the name of the time variable,
default is NULL, internal logic checks vector of possible time variable names
(t, time)}

\item{show_intervention_period}{optional logical argument (default FALSE) to
add line(s) showing the start and/or end of the intervention}

\item{intervention_period}{optional argument (default NULL) to add a vector
(type: numeric) with the start and end time of an intervention}
}
\value{
ggplot2 object of compartmental model output
}
\description{
Plots stochastic compartmental model output using \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\examples{
\dontrun{
modelout <- run_sir_stochastic(
  0.001, 0.1, 499, 100, 0, 500, 100,
  c("beta*s*i", "gamma*i"),
  matrix(c(-1, 0, +1, -1, 0, +1),
    nrow = 3, byrow = TRUE
  ), 10
)
plot_stoch_model(modelout)

modelout2 <- run_sir_stochastic_tau(
  0.00001, 0.1, 1e05 - 1, 1, 0, 1e05, 100,
  list(c(s = -1, i = 1), c(i = -1, r = 1)),
  10
)
plot_stoch_model(modelout2)
}
}
