% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_row.R
\name{process_row}
\alias{process_row}
\title{Makes a tibble row for a given interaction}
\usage{
process_row(
  tblprocesses,
  beforestates,
  afterstates,
  rate,
  metapopulation = "",
  environment = "",
  name = "",
  namegroup = "",
  beforestategroup = "",
  afterstategroup = ""
)
}
\arguments{
\item{tblprocesses}{appears unnecessary}

\item{beforestates}{vector of state names}

\item{afterstates}{vector of statenames}

\item{rate}{numeric or character for
interaction rate}

\item{metapopulation}{metapopulation it occurs,
default is "" which specifies all metapopulations}

\item{environment}{environments it occurs,
default is "" which specifies all environments}

\item{name}{name of process
default is ""}

\item{namegroup}{type of process, rename
default is ""}

\item{beforestategroup}{specifies which state groups this
interaction applies to
default is "" which specifies all groups}

\item{afterstategroup}{specifies which state groups this
interaction applies to
default is "" which specifies all groups}
}
\value{
tibble row with input values
}
\description{
Seems unnecessary, consider refactoring
}
