% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove1star.R
\name{remove1star}
\alias{remove1star}
\title{Simplifies math equations}
\usage{
remove1star(string)
}
\arguments{
\item{string}{Character vector to remove "1*" substring}
}
\value{
string with "1*" removed
}
\description{
Remove "1*" from string only if no other numbers or decimals precede 1
}
