% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceglobal.R
\name{replaceglobal}
\alias{replaceglobal}
\title{Replace empty strings in a dataframe column with a given vector}
\usage{
replaceglobal(currtbl, replacecol, replacevec)
}
\arguments{
\item{currtbl}{A dataframe/tibble with columns}

\item{replacecol}{tibble name to replace "" values}

\item{replacevec}{character vector to replace with}
}
\value{
tibble with specified column values ""
replaced with list of the vector to facilitate unnesting.
}
\description{
Replace empty strings in a dataframe column with a given vector
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
