% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_incidence.R
\name{report_incidence}
\alias{report_incidence}
\title{Aggregate incidence at reporting times}
\usage{
report_incidence(t_incident, t_report, cum_incidence = NA, t0 = 0, inc0 = 0)
}
\arguments{
\item{t_incident}{numeric vector of times when incidence changes}

\item{t_report}{numeric vector of reporting times to get aggregate incidence}

\item{cum_incidence}{numeric vector of cumulative incidence
matching t_incident
defaults to NA which treats t_incident as times of single incidence cases}

\item{t0}{optional numeric for initial reporting time
default is 0 meaning the first box contains all incident times between 0 and
the first element of t_report}

\item{inc0}{optional numeric for cumulative incidence reporting time
default is 0 meaning the first box contains all incident times between 0 and
the first element of t_report}
}
\value{
numeric vector of new incidence at reporting times
}
\description{
Aggregate incidence at reporting times
}
\seealso{
Other calculations:
\code{\link{calculate_initial_growth_rate}()},
\code{\link{odin_tot_pop}()},
\code{\link{trackincidence_byfeature}()},
\code{\link{trackincidence_byname}()}
}
\concept{calculations}
