% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_adaptivetau.R
\name{run_adaptivetau}
\alias{run_adaptivetau}
\title{Wrapper to run adaptivetau functions \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} or
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}} from compiled model and formatted inputs}
\usage{
run_adaptivetau(
  x0,
  compiledmodel,
  rate_func = NULL,
  parameters,
  t,
  method = "exact"
)
}
\arguments{
\item{x0}{the starting values for populations in each compartment}

\item{compiledmodel}{compiled model as a list, output from compilemodel()}

\item{rate_func}{function to calculate rates, defaults to NULL which uses
\code{\link[=generalized_rates]{generalized_rates()}}}

\item{parameters}{vector of parameter values}

\item{t}{number of time steps}

\item{method}{method to use in adaptivetau, defaults to "exact". Other option
is "adaptivetau"}
}
\value{
list of dataframes object of the data part of GillespieSSA object
containing the time and states of the simulation. Number of elements should
match the number of simulations
}
\description{
Wrapper to run adaptivetau functions \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} or
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}} from compiled model and formatted inputs
}
\seealso{
Other simulation:
\code{\link{gillespie}()},
\code{\link{wrap_adaptivetau}()},
\code{\link{wrap_gillespie}()},
\code{\link{wrap_ode}()},
\code{\link{wrap_odin}()}
}
\concept{simulation}
