% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gen_stochastic.r
\name{run_gen_stochastic}
\alias{run_gen_stochastic}
\title{Run generalized stochastic compartmental model}
\usage{
run_gen_stochastic(
  parms_vec,
  propensity_fns,
  init_vals,
  n_timesteps,
  change_matrix,
  n_sims,
  intervention_start_time = NULL,
  intervention_end_time = NULL,
  modifier_matrix = NULL
)
}
\arguments{
\item{parms_vec}{vector of parameter values}

\item{propensity_fns}{the transition rate equations}

\item{init_vals}{the starting values for populations in each compartment}

\item{n_timesteps}{number of time steps}

\item{change_matrix}{matrix that gives transition information between states}

\item{n_sims}{number of simulations desired}

\item{intervention_start_time}{time at which intervention will start}

\item{intervention_end_time}{time at which intervention will end}

\item{modifier_matrix}{value to be multiplied against transition matrix}
}
\value{
list of dataframes object of the data part of GillespieSSA object
containing the time and states of the simulation. Number of elements should
match the number of simulations
}
\description{
Runs basic stochastic compartmental model with an intervention using
\code{\link[GillespieSSA:ssa]{GillespieSSA::ssa()}} given a user specified number and names of
compartments, transition matrix between compartments (change matrix),
disease-related parameters, number of time steps, initial starting
conditions for each compartment (state), and the start and end time for the
intervention. No seed is set in the function so results are expected to vary
each time the function is run. Parameter values must be a vector, as
required by GillespieSSA, and ssa() function returns given compartment names.
}
\examples{
\dontrun{
initial_vals <- c(s = 990, e = 100, i1 = 10, i2 = 0, r = 0)
change_mtx <- matrix(
  c(
    -1, 0, 0, 0, # S -> E
    +1, -1, 0, 0, # E -> I1
    0, +1, -1, 0, # I1 -> I2
    0, 0, +1, -1, # I2 -> R
    0, 0, 0, +1
  ),
  nrow = length(initial_vals),
  byrow = TRUE,
  dimnames = list(names(initial_vals), NULL)
)
# Modifier matrix to reduce beta by half during the intervention period
modifier_matrix <- matrix(c(0.5),
  nrow = 1,
  dimnames = list(c("beta"), NULL)
)
results <- run_gen_stochastic_int(
  parms_vec = c(
    beta = 0.001, sigma = 0.1,
    alpha = 0.05, gamma = 0.1
  ),
  propensity_fns = c(
    "beta * s * i1",
    "sigma * e",
    "alpha * i1",
    "gamma * i2"
  ),
  init_vals = initial_vals,
  n_timesteps = 100,
  change_matrix = change_mtx,
  n_sims = 10,
  intervention_start_time = 20,
  intervention_end_time = 40,
  modifier_matrix = modifier_matrix
)
}
}
