% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gen_stochastic_tau_int.r
\name{run_gen_stochastic_tau_int}
\alias{run_gen_stochastic_tau_int}
\title{Runs generalized stochastic compartmental model with tau leaping and
intervention}
\usage{
run_gen_stochastic_tau_int(
  init_vals,
  transitions,
  rate_func,
  parms_vec,
  n_timesteps,
  n,
  n_sims,
  intervention_start_time = NULL,
  intervention_end_time = NULL,
  modifier_vec = NULL
)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}

\item{transitions}{list object with transition information between states}

\item{rate_func}{function containing the transition rate factory, e.g.
\code{\link[=generalized_rates]{generalized_rates()}} where the user
can define rate_eqns to feed into the function (see example)}

\item{parms_vec}{vector of parameter values}

\item{n_timesteps}{number of time steps}

\item{n}{population size}

\item{n_sims}{number of simulations to run}

\item{intervention_start_time}{time at which intervention will start}

\item{intervention_end_time}{time at which intervention will end}

\item{modifier_vec}{value to be multiplied against parameter vector}
}
\value{
data frame object of the time and states of the simulation
}
\description{
Runs basic stochastic SIR model using \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} given
initial conditions (states), transitions, rate function and rate
function equations, parameters, number of timesteps, population size, and
number of simulations to run.
No seed is set in the function so results are expected to vary each time
the function is run.
}
\examples{
\dontrun{
init_vals <- c(s = 99999, i = 1, r = 0)
transitions <- list(c(s = -1, i = 1), c(i = -1, r = 1))
parms_vec <- c(beta = 0.00001, gamma = 0.1, n = 1e05)
n_timesteps <- 100
n_sims <- 10
rate_eqns <- list(
  dplyr::expr(beta * s * i), # /n removed for now, will need to explore
  dplyr::expr(gamma * i)
)
out <- run_gen_stochastic_tau_int(
  init_vals, # ssa.exact requires this to be a vector, not a named list
  transitions,
  generalized_rates(rate_eqns),
  parms_vec,
  n_timesteps,
  parms_vec[["n"]],
  n_sims,
  intervention_start_time = 15,
  intervention_end_time = 20,
  modifier_vec = c(5)
)
plot_stoch_model(out, time_var = "time")
}
}
