% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sir.R
\name{run_sir}
\alias{run_sir}
\title{Runs SIR compartmental model}
\usage{
run_sir(init, time, parms)
}
\arguments{
\item{init}{vector of initial values of s0 initial count of susceptibles,
i0 initial count of infecteds, and r0 initial count of recovereds}

\item{time}{vector (type: double) containing time steps to run the model over}

\item{parms}{vector with named parameters to include: beta transmission rate,
and gamma recovery rate}
}
\value{
data frame with columns for time and each state (compartment)
}
\description{
Runs basic SIR model using \code{\link[deSolve:ode]{deSolve::ode()}} given parameters,
times, and starting conditions. Performs basic input validation via
\link{validate_sir_input}
}
\examples{
\dontrun{
run_sir(
  init = c(s = 1e05 - 1, i = 1, r = 0),
  time = seq(0.1, 100, by = 0.1),
  parms = c(beta = 0.00001, gamma = 0.1)
)
}
}
