% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sir_intervention.R
\name{run_sir_intervention}
\alias{run_sir_intervention}
\title{Runs SIR compartmental model in context of an intervention}
\usage{
run_sir_intervention(init, time, parms)
}
\arguments{
\item{init}{vector of named numeric initial values of s initial count of
susceptibles, i initial count of infecteds, and r initial count of recovereds}

\item{time}{vector (type: double) containing time steps to run the model over}

\item{parms}{vector with named parameters to include: beta transmission rate,
and gamma recovery rate, intervention_start_time time at which
intervention will start, intervention_end_time time at which intervention
will end, and the intervention_impact value to be multiplied against beta
transmission parameter to adjust transmission during the intervention}
}
\value{
data frame with columns for time and each state (compartment)
}
\description{
Runs a basic SIR model using \code{\link[deSolve:ode]{deSolve::ode()}} given parameters,
times, and starting conditions and applies an intervention impacting
transmission during a specific time range. This references
\code{\link[=sir_deriv_intervention]{sir_deriv_intervention()}} which uses explicit frequency
dependent transmission.
}
\examples{
\dontrun{
run_sir_intervention(
  init = c(s = 1e05 - 1, i = 1, r = 0),
  time = seq(0.1, 100, by = 0.1),
  parms = list(
    beta = 0.5, gamma = 0.1,
    intervention_start_time = 20,
    intervention_end_time = 30,
    intervention_impact = 0.3
  )
)
}
}
