% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sir_stochastic.R
\name{run_sir_stochastic}
\alias{run_sir_stochastic}
\title{Runs SIR stochastic compartmental model}
\usage{
run_sir_stochastic(
  beta,
  gamma,
  s0,
  i0,
  r0,
  n_pop,
  n_timesteps,
  propensity_fns,
  change_matrix,
  n_sims
)
}
\arguments{
\item{beta}{transmission rate}

\item{gamma}{recovery rate}

\item{s0}{initial count of susceptibles}

\item{i0}{initial count of infecteds}

\item{r0}{initial count of recovereds}

\item{n_pop}{population size}

\item{n_timesteps}{number of time steps}

\item{propensity_fns}{provide transition equations as a vector of equations}

\item{change_matrix}{matrix that gives transition information between states}

\item{n_sims}{number of simulations desired}
}
\value{
list of dataframes of the data part of GillespieSSA object
containing the time and states of the simulation. Number of elements should
match the number of simulations
}
\description{
Runs basic stochastic SIR model using \code{\link[GillespieSSA:ssa.d]{GillespieSSA::ssa.d()}} given
parameters, initial conditions, number of timesteps, and state change
matrix conditions. No seed is set in the function so results are expected
to vary each time the function is run.
}
\examples{
\dontrun{
run_sir_stochastic(
  0.001, 0.1, 499, 100, 0, 500, 100,
  c("beta*s*i", "gamma*i"),
  matrix(c(-1, 0, +1, -1, 0, +1),
    nrow = 3, byrow = TRUE
  ), 10
)
}
}
