% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sir_stochastic_tau.R
\name{run_sir_stochastic_tau}
\alias{run_sir_stochastic_tau}
\title{Runs SIR stochastic compartmental model with tau leaping}
\usage{
run_sir_stochastic_tau(
  beta,
  gamma,
  s0,
  i0,
  r0,
  n_pop,
  n_timesteps,
  transitions,
  n_sims
)
}
\arguments{
\item{beta}{transmission rate}

\item{gamma}{recovery rate}

\item{s0}{initial count of susceptibles}

\item{i0}{initial count of infecteds}

\item{r0}{initial count of recovereds}

\item{n_pop}{population size}

\item{n_timesteps}{number of time steps}

\item{transitions}{list object with transition information between states}

\item{n_sims}{number of simulations to run}
}
\value{
data frame object of the time and states of the simulation
}
\description{
Runs basic stochastic SIR model using \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} given
parameters, initial conditions, number of timesteps, and state change matrix.
No seed is set in the function so results are expected to vary each time
the function is run.
}
\examples{
\dontrun{
run_sir_stochastic_tau(
  0.00001, 0.1, 1e05 - 1, 1, 0, 1e05, 100,
  list(c(s = -1, i = 1), c(i = -1, r = 1)),
  10
)
}
}
