% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleprocesstblbygroups.R
\name{scaleprocesstblbygroups}
\alias{scaleprocesstblbygroups}
\title{Scales processes by grouping}
\usage{
scaleprocesstblbygroups(tblprocess, tblgroups, scalecol)
}
\arguments{
\item{tblprocess}{table of processes to be scaled as generated by compiler}

\item{tblgroups}{table of grouping process for scaling}

\item{scalecol}{column name to scale processes by type (e.g., transition,
migration, interaction)}
}
\value{
named vector with immediate names
}
\description{
For use in compiler
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
