% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_x0_byname.R
\name{set_x0_byname}
\alias{set_x0_byname}
\title{Generate initial conditions specifying value
for state names}
\usage{
set_x0_byname(tblx0, inputvec)
}
\arguments{
\item{tblx0}{tibble with "updatedstate" column
of state names}

\item{inputvec}{named list with state names as names and
values for population abundances}
}
\value{
Same tibble with "X0" column added with values
}
\description{
Generate initial conditions specifying value
for state names
}
