% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpopsize_byfeature.R
\name{setpopsize_byfeature}
\alias{setpopsize_byfeature}
\title{Set popsize for states that match input features}
\usage{
setpopsize_byfeature(
  tblpopsize,
  newpopsize,
  basestates = NA,
  groupnames = NA,
  metapopulation = NA,
  chains = NA
)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{newpopsize}{value to make all popsizes that match input features}

\item{basestates}{list/vector of basestates to match
default is NA which designates all basestates}

\item{groupnames}{named list/vector of groups to match. Names correspond to
grouptypes and the values correspond to groupnames. If names are not given
then grouptypes are automatically identified using tblpopsize.
default is NA which designates all groups}

\item{metapopulation}{list/vector of metapopulation names to match
default is NA which designates all metapopulations}

\item{chains}{list/vector of step numbers for chains to match. Names
correspond to chainid columns. If no chainid is given then it is
automatically identified from tblpopsize
default is NA which designates all chain steps}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Match features based on input. Match using OR logic within each input
list/vector of basestates, groupnames, metapopulations, and chains. While
match using AND logic between input values.
}
