% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpopsize_byname.R
\name{setpopsize_byname}
\alias{setpopsize_byname}
\title{Set popsize for named states}
\usage{
setpopsize_byname(tblpopsize, namevec)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{namevec}{named list/vector vector specifying updatedstate popsize
values to set. Does not impact unnamed states.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Set popsize for named states
}
