% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_deriv.R
\name{sir_deriv}
\alias{sir_deriv}
\title{Calculates SIR compartment states}
\usage{
sir_deriv(time, state, parms)
}
\arguments{
\item{time}{time step (type: double) the model is currently running on}

\item{state}{vector of initial values of s0 initial count of susceptibles,
i0 initial count of infecteds, and r0 initial count of recovereds}

\item{parms}{vector with named parameters to include: beta transmission rate,
and gamma recovery rate}
}
\value{
list with elements for each state (compartment)
}
\description{
Contains basic SIR model equations given parameters, times, and starting
conditions. The flow from susceptible to infectious is explicit
frequency dependent transmission.
}
\examples{
\dontrun{
sir_deriv(
  time = 1,
  state = c(s = 1e05 - 1, i = 1, r = 0),
  parms = c(beta = 0.5, gamma = 0.1)
)
}
}
