% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_deriv_intervention.R
\name{sir_deriv_intervention}
\alias{sir_deriv_intervention}
\title{Calculates SIR compartment states in context of an intervention}
\usage{
sir_deriv_intervention(time, state, parms)
}
\arguments{
\item{time}{time step (type: double) the model is currently running on}

\item{state}{vector of initial values of s0 initial count of susceptibles,
i0 initial count of infecteds, and r0 initial count of recovereds}

\item{parms}{vector with named parameters to include beta transmission rate,
and gamma recovery rate}
}
\value{
list with elements for each state (compartment) and effective beta
}
\description{
Contains basic SIR model equations given parameters, times, and starting
conditions with a calculation of the effective beta given an intervention.
The flow from susceptible to infectious is explicit frequency dependent
transmission.
}
\examples{
\dontrun{
sir_deriv_intervention(
  times = seq(0.1, 100, by = 0.1),
  state = c(s = 1e05 - 1, i = 1, r = 0),
  parms = c(
    beta = 0.5, gamma = 0.1,
    intervention_start_time = 10,
    intervention_end_time = 20,
    intervention_impact = 0.3
  )
)
}
}
