% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_rates.R
\name{sir_rates}
\alias{sir_rates}
\title{Calculates SIR flow rates}
\usage{
sir_rates(init_values, params, tf)
}
\arguments{
\item{init_values}{vector of initial values of s0 initial count of
susceptibles, #' i0 initial count of infecteds, and r0 initial count of
recovereds}

\item{params}{vector with beta transmission rate, and gamma recovery rate}

\item{tf}{vector of time steps to run the model over}
}
\value{
named numeric vector for each state (compartment)
}
\description{
Runs basic SIR model given parameters, times, and starting
conditions and returns the flows (rates of change) between comparments
for use in stochastic modeling \code{\link[=run_sir_stochastic]{run_sir_stochastic()}}
}
\examples{
\dontrun{
sir_rates(
  init_values = c(s = 1e05 - 1, i = 1, r = 0),
  params = c(beta = 0.00001, gamma = 0.1),
  tf = seq(0.1, 100, by = 0.1)
)
}
}
