% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split2typecols.R
\name{split2typecols}
\alias{split2typecols}
\title{Expand table to include group name combinations}
\usage{
split2typecols(currtbl, combinelist, listgroupcombinations)
}
\arguments{
\item{currtbl}{tibble to expand}

\item{combinelist}{list with character vector values
specifying grouptypes to make group combinations}

\item{listgroupcombinations}{the combined
groupnames based on group types}
}
\value{
tibble
}
\description{
Helper function simplifying combining group names
}
