% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitnamedlist.R
\name{splitnamedlist}
\alias{splitnamedlist}
\title{Slices named list of vectors by combining vector
elements by index in separate named lists}
\usage{
splitnamedlist(namedlist)
}
\arguments{
\item{namedlist}{named list of vectors}
}
\value{
list of named lists of 1-element vectors
}
\description{
Vectors must be same length, use matchlength.R to coerce.
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
