% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitratebynames.R
\name{splitratebynames}
\alias{splitratebynames}
\title{Replaces substrings in string}
\usage{
splitratebynames(currrate, oldnames, newnames)
}
\arguments{
\item{currrate}{String to replace substrings}

\item{oldnames}{substrings to replace (commonly basestates)}

\item{newnames}{Replacing substrings in same order
as oldnames (commonly updated state names)}
}
\value{
Full string with replaced substrings
}
\description{
Useful when updating the basestate in a
rate with the actual state name
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
