% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2petercolumn.R
\name{states2petercolumn}
\alias{states2petercolumn}
\title{Generate column for petersen matrix for a process}
\usage{
states2petercolumn(statesdown, statesup, statelist)
}
\arguments{
\item{statesdown}{State names to decrease in value
(by one). Repeat values to decrease by more than one}

\item{statesup}{State names to increase in value
(by one). Repeat values to increase by more than one}

\item{statelist}{character vector of state names
that specifies order in output matrix}
}
\value{
Column vector (as a matrix) with length(statelist) rows and
values the total integer changes to population.
}
\description{
Counts occurrences of values in input vectors and combine to get total
increase or decrease of a population due to a process
}
