% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizerate.R
\name{tokenizerate}
\alias{tokenizerate}
\title{Used to isolate user input state and parameter
names within mathematical functions}
\usage{
tokenizerate(ratestr)
}
\arguments{
\item{ratestr}{Character string}
}
\value{
Character vector where each element
are the substrings separated by common
mathematical operations.
}
\description{
Used to isolate user input state and parameter
names within mathematical functions
}
