% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackincidence_byfeature.R
\name{trackincidence_byfeature}
\alias{trackincidence_byfeature}
\title{Update compiled model with incident populations aggregated by feature}
\usage{
trackincidence_byfeature(
  compiledmodel,
  basestates = NA,
  groupnames = NA,
  metapopulation = NA,
  chains = NA,
  trackname = NA
)
}
\arguments{
\item{compiledmodel}{list of instructions for piping |>}

\item{basestates}{list/vector of basestates to match
default is NA which designates all basestates}

\item{groupnames}{named list/vector of groups to match. grouptypes are names
and groupnames are values.
default is NA which designates all groups}

\item{metapopulation}{list/vector of metapopulation names to match
default is NA which designates all metapopulations}

\item{chains}{list/vector of step numbers for chains to match. Names
correspond to chainid columns
default is NA which designates all chain steps}

\item{trackname}{character vector to name tracked population}
}
\value{
updated compiled model with added populations
}
\description{
Update compiled model with incident populations aggregated by feature
}
\seealso{
Other calculations:
\code{\link{calculate_initial_growth_rate}()},
\code{\link{odin_tot_pop}()},
\code{\link{report_incidence}()},
\code{\link{trackincidence_byname}()}
}
\concept{calculations}
