% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackincidence_byname.R
\name{trackincidence_byname}
\alias{trackincidence_byname}
\title{Update compiled model with incident populations aggregated by name}
\usage{
trackincidence_byname(compiledmodel, namevec, trackname = NA)
}
\arguments{
\item{compiledmodel}{list of instructions for piping |>}

\item{namevec}{vector of updated state names to aggregate}

\item{trackname}{character vector to name tracked population}
}
\value{
updated compiled model with added populations
}
\description{
Update compiled model with incident populations aggregated by name
}
