% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Updates parameters due to modifications during model run}
\usage{
update_parameters(current_parms_vec, modifier_matrix)
}
\arguments{
\item{current_parms_vec}{vector of current model parameter values}

\item{modifier_matrix}{value to be multiplied against transition matrix}
}
\value{
vector of modified parameter values
}
\description{
This helper function updates model parameters as a result of an intervention
or other modification
}
\examples{
\dontrun{
updated_parms_vec <- update_parameters(current_parms_vec, modifier_matrix)
}
}
