% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gen_determ_input.R
\name{validate_gen_determ_input}
\alias{validate_gen_determ_input}
\title{Checks validity of input to generalized compartmental model}
\usage{
validate_gen_determ_input(init_vals, times, comp_names, trans_matrix)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}

\item{times}{vector of time steps to run the model over}

\item{comp_names}{names of the compartments}

\item{trans_matrix}{matrix of transition rates between each of the}
}
\value{
stops with information or indicates Input Checks Passed
}
\description{
Helper function that runs checks on generalized compartmental model input
parameters, times, and initial conditions
}
\examples{
\dontrun{
validate_gen_determ_input(
  init_vals = c(1e05 - 1, 10, 5, 1, 0),
  times = seq(0.1, 100, by = 0.1),
  comp_names = c("A", "B", "C", "D", "E"),
  trans_matrix <- matrix(
  c(
      -0.1, 0.1, 0, 0, 0,
      0, -0.1, 0.1, 0, 0,
      0, 0, -0.1, 0.1, 0,
      0, 0, 0, -0.1, 0.1,
      0.1, 0, 0, 0, -0.1
    ),
    nrow = length(c("A", "B", "C", "D", "E")),
    byrow = TRUE,
    dimnames = list(c("A", "B", "C", "D", "E"), c("A", "B", "C", "D", "E"))
  )
)
}
}
