% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gen_stoch_input.r
\name{validate_gen_stoch_input}
\alias{validate_gen_stoch_input}
\title{Checks validity of input to stochastic compartmental model}
\usage{
validate_gen_stoch_input(
  parms_vec,
  init_vals,
  n_timesteps,
  method = NULL,
  propensity_fns = NULL,
  change_matrix = NULL,
  transitions = NULL,
  n_sims,
  intervention_start_time = NULL,
  intervention_end_time = NULL,
  modifier = NULL
)
}
\arguments{
\item{parms_vec}{vector of parameter values}

\item{init_vals}{the starting values for populations in each compartment}

\item{n_timesteps}{number of time steps}

\item{method}{character name of stochastic method used, default NULL. Example
options include "GillespieSSA" or "adaptivetau"}

\item{propensity_fns}{provide transition equations as a vector of equations,
default NULL}

\item{change_matrix}{matrix that gives transition information between states,
default NULL}

\item{transitions}{list object with transition information between states,
default NULL}

\item{n_sims}{number of simulations desired}

\item{intervention_start_time}{time at which intervention will start}

\item{intervention_end_time}{time at which intervention will end}

\item{modifier}{single value, matrix of values, or user supplied function to
to be multiplied against transition matrix}
}
\value{
stops with information or indicates Input Checks Passed
}
\description{
Helper function that runs checks on generalized compartmental model input
parameters, times, and initial conditions
}
