% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sir_input.R
\name{validate_sir_input}
\alias{validate_sir_input}
\title{Checks validity of input to SIR model}
\usage{
validate_sir_input(init, time, parms, intervention = FALSE)
}
\arguments{
\item{init}{vector of initial values of s0 initial count of susceptibles,
i0 initial count of infecteds, and r0 initial count of recovereds}

\item{time}{vector (type: double) containing time steps to run the model over}

\item{parms}{vector with beta transmission rate, and gamma recovery rate}

\item{intervention}{logical (default FALSE) to indicate if testing input
parameters related to an intervention need to be checked}
}
\value{
stops with information or indicates "Input Checks Passed"
}
\description{
Helper function that runs checks on basic SIR model input parameters, times,
and initial conditions
}
\examples{
\dontrun{
validate_sir_input(
  init = c(1e05 - 1, 1, 0),
  times = seq(0.1, 100, by = 0.1),
  parms = c(0.00001, 0.1)
)
}
}
