% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sir_stoch_input.R
\name{validate_sir_stoch_input}
\alias{validate_sir_stoch_input}
\title{Checks validity of input to stochastic SIR model}
\usage{
validate_sir_stoch_input(
  beta,
  gamma,
  s0,
  i0,
  r0,
  n_pop,
  n_timesteps,
  method = NULL,
  propensity_fns = NULL,
  change_matrix = NULL,
  transitions = NULL,
  n_sims
)
}
\arguments{
\item{beta}{transmission rate}

\item{gamma}{recovery rate}

\item{s0}{initial count of susceptibles}

\item{i0}{initial count of infecteds}

\item{r0}{initial count of recovereds}

\item{n_pop}{population size}

\item{n_timesteps}{number of time steps}

\item{method}{character name of stochastic method used, default NULL. Example
options include "GillespieSSA" or "adaptivetau"}

\item{propensity_fns}{provide transition equations as a vector of equations,
default NULL}

\item{change_matrix}{matrix that gives transition information between states,
default NULL}

\item{transitions}{list object with transition information between states,
default NULL}

\item{n_sims}{number of simulations desired}
}
\value{
stops with information or indicates Input Checks Passed
}
\description{
Helper function that runs checks on basic SIR stochastic model input
parameters, times, and initial conditions
}
\examples{
\dontrun{
validate_sir_stoch_input(
  0.001, 0.1, 499, 100, 0, 500, 100,
  method = "GillespieSSA",
  c("beta*s*i", "gamma*i"),
  matrix(c(-1, 0, +1, -1, 0, +1),
    nrow = 3, byrow = TRUE
  ), 10
)

validate_sir_stoch_input(
  0.001, 0.1, 499, 100, 0, 500, 100,
  method = "adaptivetau",
  list(c(s = -1, i = 1), c(i = -1, r = 1)),
  10
)
}
}
