% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_adaptivetau.R
\name{wrap_adaptivetau}
\alias{wrap_adaptivetau}
\title{Wrapper to run adaptivetau functions \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} or
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}} from compiled model and formatted inputs}
\usage{
wrap_adaptivetau(
  init_vals,
  compiledmodel,
  rate_func = NULL,
  parameters,
  n_timesteps,
  n_sims,
  method = "exact"
)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}

\item{compiledmodel}{compiled model as a list, output from compilemodel()}

\item{rate_func}{function to calculate rates, defaults to NULL which uses
\code{\link[=generalized_rates]{generalized_rates()}}}

\item{parameters}{vector of parameter values}

\item{n_timesteps}{number of time steps}

\item{n_sims}{number of simulations desired}

\item{method}{method to use in adaptivetau, defaults to "exact". Other option
is "adaptivetau"}
}
\value{
list of data frames object of the data part of adaptivetau object
containing the time and states of the simulation. Number of elements should
match the number of simulations
}
\description{
Wrapper to run adaptivetau functions \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} or
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}} from compiled model and formatted inputs
}
\examples{
\dontrun{
# SIR
# Note this example does not name each argument so rate_func = NULL is needed
# for R to recognize the correct arguments in order
base_states <- c("S", "I", "R")
sir <- define_states(base_states) |>
  add_infection("I", "S", "I", "beta") |>
  add_transition("I", "R", "tau")
sircompiled <- compilemodel(sir)
test1 <- wrap_adaptivetau(
  c("S" = 999, "I" = 1, "R" = 0),
  sircompiled,
  rate_func = NULL,
  c(beta = 2, tau = 1),
  25,
  10
)

# SIR with tau leaping
base_states <- c("S", "I", "R")
sir <- define_states(base_states) |>
  add_infection("I", "S", "I", "beta") |>
  add_transition("I", "R", "tau")
sircompiled <- compilemodel(sir)
test1 <- wrap_adaptivetau(
  c("S" = 999, "I" = 1, "R" = 0),
  sircompiled,
  rate_func = NULL,
  c(beta = 2, tau = 1),
  25,
  10,
  "adaptivetau"
)

# SEIR
# Note: this example names each argument so rate_func can be left out
base_states <- c("S", "E", "I", "R")
seir <- define_states(base_states) |>
  add_infection("I", "S", "E", "beta") |>
  add_transition("I", "R", "tau") |>
  add_transition("E", "I", "taue")
seircompiled <- compilemodel(seir)
test2 <- wrap_adaptivetau(
  c("S" = 999, "I" = 1, "R" = 0, "E" = 0),
  seircompiled,
  rate_func = NULL,
  c(beta = 2, tau = 1, taue = .5),
  25,
  10
)
}
}
\seealso{
Other simulation:
\code{\link{gillespie}()},
\code{\link{run_adaptivetau}()},
\code{\link{wrap_gillespie}()},
\code{\link{wrap_ode}()},
\code{\link{wrap_odin}()}
}
\concept{simulation}
