% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_gillespie.R
\name{wrap_gillespie}
\alias{wrap_gillespie}
\title{Wrapper to run Gillespie algorithm with \code{\link[GillespieSSA:ssa]{GillespieSSA::ssa()}} from compiled
model and formatted inputs}
\usage{
wrap_gillespie(init_vals, compiledmodel, parameters, n_timesteps, n_sims)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}

\item{compiledmodel}{compiled model as a list, output from compilemodel()}

\item{parameters}{vector of parameter values}

\item{n_timesteps}{number of time steps}

\item{n_sims}{number of simulations desired}
}
\value{
list of data frames object of the data part of GillespieSSA object
containing the time and states of the simulation. Number of elements should
match the number of simulations
}
\description{
Wrapper to run Gillespie algorithm with \code{\link[GillespieSSA:ssa]{GillespieSSA::ssa()}} from compiled
model and formatted inputs
}
\examples{
\dontrun{
# SIR
base_states <- c("S", "I", "R")
sir <- define_states(base_states) |>
  add_infection("I", "S", "I", "beta") |>
  add_transition("I", "R", "tau")
sircompiled <- compilemodel(sir)
wrap_gillespie(
  c("S" = 999, "I" = 1, "R" = 0),
  sircompiled,
  c(beta = 2, tau = 1),
  25,
  10
)

# SEIR
base_states <- c("S", "E", "I", "R")
seir <- define_states(base_states) |>
  add_infection("I", "S", "E", "beta") |>
  add_transition("I", "R", "tau") |>
  add_transition("E", "I", "taue")
seircompiled <- compilemodel(seir)
test2 <- wrap_gillespie(
  c("S" = 999, "I" = 1, "R" = 0, "E" = 0),
  seircompiled,
  c(beta = 2, tau = 1, taue = .5),
  25,
  10
)
}
}
