% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ode.R
\name{wrap_ode}
\alias{wrap_ode}
\title{Wrapper to run \code{\link[deSolve:ode]{deSolve::ode()}} from compiled model and formatted inputs}
\usage{
wrap_ode(y, times, compiledmodel, parms)
}
\arguments{
\item{y}{the initial (state) values for the ODE system, a named vector}

\item{times}{vector of sequence of times for which output is wanted; first
value must be the initial time}

\item{compiledmodel}{compiled model as a list, output from compilemodel()}

\item{parms}{named vector of parameters passed to the ode function created by
model2desolvefunction using the compiledmodel}
}
\value{
data frame of deSolve output run on provided input
}
\description{
Wrapper to run \code{\link[deSolve:ode]{deSolve::ode()}} from compiled model and formatted inputs
}
\examples{
\dontrun{
base_states <- c("S", "I", "R")
sir <- define_states(base_states) |>
  add_infection("I", "S", "I", "beta") |>
  add_transition("I", "R", "tau")
sircompiled <- compilemodel(sir)
wrap_ode(
  c("S" = 999, "I" = 1, "R" = 0),
  seq(0, 25, by = 0.01),
  sircompiled,
  c(beta = 2, tau = 1)
)
#'
}
}
