% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_odin.R
\name{wrap_odin}
\alias{wrap_odin}
\title{Takes input from compiledmodel and additional model information,
creates model in format for odin (needs to be saved as an R file),
runs odin model, returns results in dataframe}
\usage{
wrap_odin(init_vals, compiledmodel, param_vector, n_timesteps, n_sims, t0 = 0)
}
\arguments{
\item{init_vals}{named vector of initial values for each state
Example: init_vals <- c("S" = 999, "I" = 1, "R" = 0, "E" = 0)}

\item{compiledmodel}{compiled model object
created by compModels function compilemodel()}

\item{param_vector}{vector of parameter names and values
Example: param_vector <-c("beta"= 2, "alpha" = 1, "gamma" = 1)}

\item{n_timesteps}{number of timesteps to run the model for}

\item{n_sims}{number of simulations to run}

\item{t0}{initial time for model run, default is 0}
}
\value{
dataframe of results from odin model run
}
\description{
Takes input from compiledmodel and additional model information,
creates model in format for odin (needs to be saved as an R file),
runs odin model, returns results in dataframe
}
\seealso{
Other simulation:
\code{\link{gillespie}()},
\code{\link{run_adaptivetau}()},
\code{\link{wrap_adaptivetau}()},
\code{\link{wrap_gillespie}()},
\code{\link{wrap_ode}()}
}
\concept{simulation}
