% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group.R
\name{add_group}
\alias{add_group}
\title{Stratify population into groups}
\usage{
add_group(
  peterlist,
  groupnames,
  grouptype = NA,
  basestates = "",
  metapopulation = "",
  scaleinteractions = 1,
  scaletransitions = 1,
  scalemigrations = 1,
  scaleprocessbyname = list(),
  scaleprocessbygroup = list()
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{groupnames}{character vector of groups
in a given grouptype (e.g., c("Old","Young"))}

\item{grouptype}{character vector of grouptype (e.g., c("Age"))
default is NA which automatically generates a type during
model compilation}

\item{basestates}{character vector of basestates
to stratify. default is "" which specifies all basestates}

\item{metapopulation}{chracter vector of metapopulations where
basestates are split}

\item{scaleinteractions}{character/numeric specifies
how to scale interaction rates for each group. default is 1}

\item{scaletransitions}{character/numeric specifies
how to scale transition rates for each group. default is 1.}

\item{scalemigrations}{character/numeric specifies
how to scale migration rates for each group. default is 1.}

\item{scaleprocessbyname}{named list w/ character/numeric values that
scale named processes, allows user defined flexibility}

\item{scaleprocessbygroup}{named list w/ character/numeric values that
scale processes by grouped name, allows user defined flexibility}
}
\value{
updated model instructions
}
\description{
Stratify population into groups
}
