% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_infection.R
\name{add_infection}
\alias{add_infection}
\title{add infection to model}
\usage{
add_infection(
  peterlist,
  infector,
  infectee,
  infected,
  rate,
  normlogic = TRUE,
  metapopulation = "",
  processname = NA,
  processgroup = NA,
  groupnames = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{infector}{base state that is infectious}

\item{infectee}{base state that is susceptible}

\item{infected}{base state that infectee
converts to following infectious}

\item{rate}{character/numeric value specifying
PER CAPITA rate of interaction}

\item{normlogic}{logical specifying if rate
should be divided by total population in given
metapopulation}

\item{metapopulation}{character vector of
metapopulations where this interaction occurs
default is "" which specifies all metapopulations}

\item{processname}{character value for referencing the added process
by other functions
default to NA}

\item{processgroup}{character value that groups named process to quickly
reference multiple processes by other functions
default to NA}

\item{groupnames}{specifies which stratified groups
transition
default is "" which specifies all groups.}
}
\value{
updated instruction list
}
\description{
Helper function to add specific type of interaction
}
