% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_infection.R
\name{add_infection}
\alias{add_infection}
\title{add infection to model}
\usage{
add_infection(
  peterlist,
  infector,
  infectee,
  infected,
  rate,
  normlogic = TRUE,
  environment = "",
  metapopulation = "",
  name = "infection",
  groups = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{infector}{base state that is infectious}

\item{infectee}{base state that is susceptible}

\item{infected}{base state that infectee
converts to following infectious}

\item{rate}{character/numeric value specifying
PER CAPITA rate of interaction}

\item{normlogic}{logical specifying if rate should be
divided by total population in given environment/metapopulation.
default is TRUE which scales rate by
total population (e.g., beta\emph{S}I/N)}

\item{environment}{character vector of environments where
this interaction occurs.
default is "" which specifies all environments}

\item{metapopulation}{character vector of metapopulations
where this interaction occurs.
default is "" which specifies all metapopulations}

\item{name}{character value to allow scaling by other functions.
default is "infection"}

\item{groups}{specifies grouptypes that engage in this interaction.
default is "" which specifies all types.}
}
\value{
updated instruction list
}
\description{
Helper function to add specific type of interaction
}
