% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interaction.R
\name{add_interaction}
\alias{add_interaction}
\title{add interactions to model}
\usage{
add_interaction(
  peterlist,
  states_in,
  states_out,
  rate,
  normlogic = TRUE,
  metapopulation = "",
  processname = NA,
  processgroup = NA,
  groupnames = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{states_in}{character vector of base state names
that interact}

\item{states_out}{character vector of base
state names after interactions. Assume states_in
and states_out are ordered (e.g., S,I -> I,I)}

\item{rate}{character/numeric value specifying
PER CAPITA rate of interaction}

\item{normlogic}{logical specifying if rate
should be divided by total population in given
metapopulation}

\item{metapopulation}{character vector of
metapopulations where this interaction occurs
default is "" which specifies all metapopulations}

\item{processname}{character value for referencing the added process
by other functions
default to NA}

\item{processgroup}{character value that groups named process to quickly
reference multiple processes by other functions
default to NA}

\item{groupnames}{Groups the transition applies to.
Either a character vector of groupnames or named list with grouptype names
and groupname values
default is "" which specifies all groups.}
}
\value{
updated instruction list
}
\description{
general form where users can specify states up
and down. Assumes per capita rates of interaction!
}
