% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interaction.R
\name{add_interaction}
\alias{add_interaction}
\title{add interactions to model}
\usage{
add_interaction(
  peterlist,
  states_in,
  states_out,
  rate,
  normlogic = TRUE,
  environment = "",
  metapopulation = "",
  name = "",
  groups = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{states_in}{character vector of base state names
that interact}

\item{states_out}{character vector of base
state names after interactions. Assume states_in
and states_out are ordered (e.g., S,I -> I,I)}

\item{rate}{character/numeric value specifying
PER CAPITA rate of interaction}

\item{normlogic}{logical specifying if rate
should be divided by total population in given
environment/metapopulation}

\item{environment}{character vector of environments
where this interaction occurs.
default is "" which specifies all environments}

\item{metapopulation}{character vector of
metapopulations where this interaction occurs
default is "" which specifies all metapopulations}

\item{name}{character value to allow scaling
by other functions
default is ""}

\item{groups}{specifies grouptypes the engage
in this interaction
default is "" which specifies all groups}
}
\value{
updated instruction list
}
\description{
general form where users can specify states up
and down. Assumes per capita rates of interaction!
}
