% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metapopulations.R
\name{add_metapopulations}
\alias{add_metapopulations}
\title{Adds metapopulations to model instructions}
\usage{
add_metapopulations(
  peterlist,
  metapop_names = "",
  environment_names = "",
  interactionscale = 1,
  transitionscale = 1,
  basestates = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{metapop_names}{character vector of
metapopulation names to add.
default is "" where compilation automatically
generates a placeholder name for
each add_metapopulation use}

\item{environment_names}{character vector of
environment names to add
default is ""}

\item{interactionscale}{character/numeric vector
specifying multiplicative scaling of interaction
rates in each specifed metapopulation
default is 1}

\item{transitionscale}{character/numeric vector
specifying multiplicative scaling of transition
rates in each specifed metapopulation
default is 1}

\item{basestates}{specifies which basestates can
visit metapopulation/environment
default is "" which species all basestates}
}
\value{
updated instruction list
}
\description{
Allows nested input of metapopulations and
environments. Maybe clean this up.
}
