% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_replacement.R
\name{add_replacement}
\alias{add_replacement}
\title{add replacement in model}
\usage{
add_replacement(
  peterlist,
  statesout,
  statesin,
  meantime,
  name = "replacement",
  chainlength = 1,
  chaintimescale = NA,
  percapita = TRUE,
  metapopulation = "",
  environment = "",
  groupname = "",
  grouptype = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{statesout}{character vector of state names
to transition from}

\item{statesin}{character vector of basestates to
transition to}

\item{meantime}{character/numeric value specifying
average time to go from fromstate to all tostates}

\item{name}{character value to allow scaling by
other functions
default is "replacement"}

\item{chainlength}{total steps from fromstate to
tostates (i.e., boxcars or length of pitchfork).
Alters distribution of transition times during
stochastic implementation
default is 1 for exponentially distributed wait times}

\item{chaintimescale}{character/numeric vector with
length chainlength scales rate transitioning between
steps along chains. Allows deviating transition time
distribution from a gamma distributions
default is NA which defaults to the same transition
times between all chained states}

\item{percapita}{specifies rates are given as per capita
value so the total change to population should be
scaled by the tostate
default to TRUE}

\item{metapopulation}{character vector of
metapopulation
names this transitions occurs in. default "" specifies
all metapopulations
default to "" which specifies all metapopulations}

\item{environment}{character vector of environment names
this transitions occurs in
default to "" which specifies all environments}

\item{groupname}{specifies which stratified groups transition
default to "" which specifies all groups.}

\item{grouptype}{specifies which stratified
group types transition
default to "" which specifies all grouptypes.}
}
\value{
updated instruction list
}
\description{
Helper fucntion that is same as transiton,
but allows simplifies specifying independent transitions
from multiple states to a single state.
}
