% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_travel.R
\name{add_travel}
\alias{add_travel}
\title{add travel between metapopulations in model}
\usage{
add_travel(
  peterlist,
  rates,
  frommetapopulation = "",
  tometapopulation = "",
  processname = NA,
  processgroup = NA,
  travelbasestates = "",
  travelgroups = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{rates}{character/numeric vector movement rate/s}

\item{frommetapopulation}{character vector of origin
metapopulations
default is "" which specifies all metapopulations}

\item{tometapopulation}{character vector of destination
metapopulations. default is "" which specifies all
metapopulations}

\item{processname}{string to name process}

\item{processgroup}{string to group process}

\item{travelbasestates}{character vector of basestates that
travel between metapopulations
default is "" which specifies all basestates}

\item{travelgroups}{character vector of groups that
travel between metapopulations
default is "" which specifies all basestates}
}
\value{
updated instruction list
}
\description{
add travel between metapopulations in model
}
