% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_travel.R
\name{add_travel}
\alias{add_travel}
\title{add travel between metapopulations in model}
\usage{
add_travel(
  peterlist,
  rates,
  frommetapopulation = "",
  tometapopulation = "",
  processname = NA,
  processgroup = NA,
  travelbasestates = "",
  travelgroups = ""
)
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{rates}{character/numeric vector movement rate/s}

\item{frommetapopulation}{character vector of origin
metapopulations
default is "" which specifies all metapopulations}

\item{tometapopulation}{character vector of destination
metapopulations. default is "" which specifies all
metapopulations}

\item{processname}{string to name process}

\item{processgroup}{string to group process}

\item{travelbasestates}{character vector of basestates that
travel between metapopulations
default is "" which specifies all basestates}

\item{travelgroups}{character vector of groups that
travel between metapopulations
default is "" which specifies all basestates}
}
\value{
updated instruction list
}
\description{
add travel between metapopulations in model
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
