% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpopsize_byname.R
\name{addpopsize_byname}
\alias{addpopsize_byname}
\title{Add popsize to named states
Add specified value to popsize of specified named state. Input tblpopsize
must have a numeric column called "popsize"}
\usage{
addpopsize_byname(tblpopsize, namevec)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{namevec}{named list/vector vector specifying updatedstate popsize
values to set. Does not impact unnamed states.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Add popsize to named states
Add specified value to popsize of specified named state. Input tblpopsize
must have a numeric column called "popsize"
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{filterpopsize_byfeature}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
