% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputpopsize.R
\name{check_inputpopsize}
\alias{check_inputpopsize}
\title{Ensure updatedstates input has popsize column}
\usage{
check_inputpopsize(input)
}
\arguments{
\item{input}{input to check}
}
\value{
error if missing column
}
\description{
Ensure updatedstates input has popsize column
}
