% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanscale.R
\name{cleanscale}
\alias{cleanscale}
\title{Simplify auto-generated math expressions}
\usage{
cleanscale(currtbl)
}
\arguments{
\item{currtbl}{A dataframe/tibble with columns
"scaleinteractions", "scaletransitions", "scalemigrations"}
}
\value{
tibble with updated scaleinteractions, scaletransitions, and
scalemigrations values
}
\description{
This function appends "\emph{" to rates and then converts the rates "1}"
and "*" to empty strings "".
This facilitates and simplifies specifying rates
auto-generated during model compilation
}
