% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanscale.R
\name{cleanscale}
\alias{cleanscale}
\title{Simplify auto-generated math expressions}
\usage{
cleanscale(currtbl)
}
\arguments{
\item{currtbl}{A dataframe/tibble with columns
"scaleinteractions", "scaletransitions", "scalemigrations"}
}
\value{
tibble with updated scaleinteractions, scaletransitions, and
scalemigrations values
}
\description{
This function appends "\emph{" to rates and then converts the rates "1}"
and "*" to empty strings "".
This facilitates and simplifies specifying rates
auto-generated during model compilation
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
