% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_groups.R
\name{combine_groups}
\alias{combine_groups}
\title{Combine group types}
\usage{
combine_groups(peterlist, grouptypes2combine = "")
}
\arguments{
\item{peterlist}{list of instructions for piping |>}

\item{grouptypes2combine}{character vector of grouptypes to combine}
}
\value{
updated model instructions
}
\description{
Adds instructions to add states that are combinations of a given type
}
