% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilemodel.R
\name{compilemodel}
\alias{compilemodel}
\title{Compiles model instructions to generate dynamical model
(updated state names, petersen matrix, and process rates)}
\usage{
compilemodel(inputpeter)
}
\arguments{
\item{inputpeter}{model instructions list.}
}
\value{
list with updated instructions $modelinstructions to
aid user in verifying model and $modeloustrcutions list
with petersen matrix, process rates, and updated state names
}
\description{
Compiles model instructions to generate dynamical model
(updated state names, petersen matrix, and process rates)
}
