% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_initialstate.R
\name{define_initialstate}
\alias{define_initialstate}
\title{Set initial population states for compiled
model using base states}
\usage{
define_initialstate(outlist, namedvector = c())
}
\arguments{
\item{outlist}{compiled model output}

\item{namedvector}{numeric vector specifying inital populations
for named basestates
default is c() which specifies all 0 population
for each basestate}
}
\value{
tibble with columns of updated state names
(updatedstates) and initial conditions (X0)
}
\description{
Sets globally based on base states. Update for more flexibility.
}
