% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_states.R
\name{define_states}
\alias{define_states}
\title{Initializes model instructions with given base states}
\usage{
define_states(base_state, environment_name = "", metapop_name = "")
}
\arguments{
\item{base_state}{character vector of base states.}

\item{environment_name}{character vector environment names
default is ""}

\item{metapop_name}{character vector of metapopulation names
default is ""}
}
\value{
list of model instructions
}
\description{
Initializes model instructions with given base states
}
