% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty2states.R
\name{empty2states}
\alias{empty2states}
\title{Expands instructions to apply globally}
\usage{
empty2states(x, states)
}
\arguments{
\item{x}{character vector.}

\item{states}{list of values to replace}
}
\value{
character vector with replaced values
}
\description{
Replace "" with given list. Write unit
test and check where used for difference
between vectors and list.
}
