% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty2states.R
\name{empty2states}
\alias{empty2states}
\title{Replace "" in a vector with a given vector}
\usage{
empty2states(x, states)
}
\arguments{
\item{x}{character vector.}

\item{states}{list of values to replace}
}
\value{
character vector with replaced values
}
\description{
Return the given x vector replacing all occurrences of "" with the given
states vector. Ordering of "" replacement is not maintained since it's a
many-to-one replacement.
}
