% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterpopsize_byfeature.R
\name{filterpopsize_byfeature}
\alias{filterpopsize_byfeature}
\title{Filter popsize table to match input features}
\usage{
filterpopsize_byfeature(
  tblpopsize,
  basestates = NA,
  groupnames = NA,
  metapopulation = NA,
  chains = NA
)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{basestates}{list/vector of basestates to match
default is NA which designates all basestates}

\item{groupnames}{named list/vector of groups to match. grouptypes are names
and groupnames are values.
default is NA which designates all groups}

\item{metapopulation}{list/vector of metapopulation names to match
default is NA which designates all metapopulations}

\item{chains}{list/vector of step numbers for chains to match. Names
correspond to chainid columns
default is NA which designates all chain steps}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Match features based on input. Match using OR logic within each input
list/vector of basestates, groupnames, metapopulations, and chains. While
match using AND logic between input values.
}
\seealso{
Other model_building:
\code{\link{add_group}()},
\code{\link{add_grouptransition}()},
\code{\link{add_infection}()},
\code{\link{add_interaction}()},
\code{\link{add_metapopulations}()},
\code{\link{add_replacement}()},
\code{\link{add_transition}()},
\code{\link{add_travel}()},
\code{\link{addpopsize_byfeature}()},
\code{\link{addpopsize_byname}()},
\code{\link{combine_groups}()},
\code{\link{compilemodel}()},
\code{\link{define_metapopulations}()},
\code{\link{generalized_rates}()},
\code{\link{make_metapopulations}()},
\code{\link{movefracpopsize_byfeature}()},
\code{\link{movefracpopsize_byname}()},
\code{\link{movepopsize_byfeature}()},
\code{\link{movepopsize_byname}()},
\code{\link{movethresholdpopsize_byfeature}()},
\code{\link{movethresholdpopsize_byname}()},
\code{\link{odin_core_eqns}()},
\code{\link{odin_draws}()},
\code{\link{odin_probs}()},
\code{\link{removepopsize_byfeature}()},
\code{\link{removepopsize_byname}()},
\code{\link{setpopsize_byfeature}()},
\code{\link{setpopsize_byname}()}
}
\concept{model_building}
