% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixinput_groupnames.R
\name{fixinput_groupnames}
\alias{fixinput_groupnames}
\title{Convert groupnames input to named list}
\usage{
fixinput_groupnames(groupnames, tblwgroupcols)
}
\arguments{
\item{groupnames}{vector or named list of groupnames. values are elements of
input table}

\item{tblwgroupcols}{table that has grouptype columns with groupnames as
values}
}
\value{
groupnames coerced into named list with respective grouptype as name
}
\description{
groupnames input can be vector or named list, depending on model complexity.
This function coerces input so that it is a named list with names matching
table columns that contain input values. The resulting consistent format
simplifies manipulating user input across functions.
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
