% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_rates.r
\name{generalized_rates}
\alias{generalized_rates}
\title{Calculates rates along model in/out flows from compartments
Helper function to calculate modified change rates for in and out flows in
a stochastic generalized compartmental model using \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}}}
\usage{
generalized_rates(rate_eqns)
}
\arguments{
\item{rate_eqns}{a list of the transition rate equations between flows,
listed in the same order as the transitions supplied to the
model runner function}
}
\value{
the rate(s) at a specific time point
}
\description{
Calculates rates along model in/out flows from compartments
Helper function to calculate modified change rates for in and out flows in
a stochastic generalized compartmental model using \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}}
}
