% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabfromtbl.R
\name{grabfromtbl}
\alias{grabfromtbl}
\title{Grabs rows that matches values in (named)list}
\usage{
grabfromtbl(namedlist, tbl2grab)
}
\arguments{
\item{namedlist}{(named)list of groups to grab. If names are given they
correspond to types}

\item{tbl2grab}{table with that has columns with values matching those in the
(named)list. If a named list is input, then some column names match the
namedlist names.}
}
\value{
table with rows that contain the named values
}
\description{
Names of input named list are column names.
}
