% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcombinations.R
\name{groupcombinations}
\alias{groupcombinations}
\title{Generates all combinations groups with different grouptypes}
\usage{
groupcombinations(combinethesetypes, tblgroup)
}
\arguments{
\item{combinethesetypes}{list of character vectors that
specify grouptypes to combine groups.}

\item{tblgroup}{tibble specifying groups in each group type}
}
\value{
tibble where each row is a combination of groups
with grouptype columns and scaled process rates
}
\description{
Combines groups based on input group types. Useful when compiling to build
updated state table.
}
