% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouptypecombinations.R
\name{grouptypecombinations}
\alias{grouptypecombinations}
\title{Combines groups by type}
\usage{
grouptypecombinations(combinethesetypes, tblgrouptype, typecolumn, groupcolumn)
}
\arguments{
\item{combinethesetypes}{character vector specifying types to combine groups
across}

\item{tblgrouptype}{tibble with given type and group columns}

\item{typecolumn}{column name for types in tblgrouptype}

\item{groupcolumn}{column name for groups in tblgrouptype}
}
\value{
tibble of all group combinations with types as columns
}
\description{
Combines groups based on input group types. Useful when compiling to build
updated state table.
}
