% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intlogic.R
\name{intlogic}
\alias{intlogic}
\title{Determine if vector has integer values}
\usage{
intlogic(vec)
}
\arguments{
\item{vec}{vector to check values}
}
\value{
logical for containing integers
}
\description{
Determine if vector has integer values
}
