% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metapopulations.R
\name{make_metapopulations}
\alias{make_metapopulations}
\title{Set metapopulations in model}
\usage{
make_metapopulations(
  metapop_names = "",
  environment_names = "",
  interactionscale = 1,
  transitionscale = 1,
  basestates = ""
)
}
\arguments{
\item{metapop_names}{vector of metapopulation names
default is ""}

\item{environment_names}{vector/list of environment names
default is ""}

\item{interactionscale}{vector same length as
input metapopulations/environment_names that
scales all interactions in given metapop/environment
default is 1}

\item{transitionscale}{vector same length as
input metapopulations/environment_names that
scales all transitions in given metapop/environment
default is 1}

\item{basestates}{specifies which states can
visit metapopulation/environment.
default is "" which specifies all basestates}
}
\value{
list combining all information.
}
\description{
This function may be deprecated, check input
output. Inputs instructions specifying metapopulations
in model. Overwrites any prior
instructions to add_metapopulation. Input can be
nested, although this functionality is
unnecessarilly complex and should be simplified.
}
