% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchlength.R
\name{matchlength}
\alias{matchlength}
\title{Convert vector to input length}
\usage{
matchlength(vec, desiredlength)
}
\arguments{
\item{vec}{character or numeric vector.}

\item{desiredlength}{output vector length}
}
\value{
lengthened vector same type as input
}
\description{
Replicates length one vector and ensures n-length
vectors are same size as given length.
Same as matchandpaste.R
}
