% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchlength_namedlist.R
\name{matchlength_namedlist}
\alias{matchlength_namedlist}
\title{Convert named list for input as tibble column}
\usage{
matchlength_namedlist(namedlist, desiredlength)
}
\arguments{
\item{namedlist}{named list to split}

\item{desiredlength}{output list length}
}
\value{
list of lists for tibble input
}
\description{
Replicates length one vector and ensures n-length
vectors are same size as given length.
}
