% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedunlist.R
\name{mixedunlist}
\alias{mixedunlist}
\title{Unlist named list to get immediate names of each element}
\usage{
mixedunlist(mixedlist)
}
\arguments{
\item{mixedlist}{list of named vectors}
}
\value{
named vector with immediate names
}
\description{
Useful when getting values associated to group names from a list that has
both groupnames and typenames
}
