% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeloutput2odeinput.R
\name{modeloutput2odeinput}
\alias{modeloutput2odeinput}
\title{Generate deSolve input from compiled model}
\usage{
modeloutput2odeinput(outlist)
}
\arguments{
\item{outlist}{compiled model}
}
\value{
character vector with strings corresponding
to instantaneous population change as values.
}
\description{
Generate deSolve input from compiled model
}
\seealso{
Other conversions:
\code{\link{cuminc2inc}()},
\code{\link{delaytime_byfunction}()},
\code{\link{delaytime_exp}()},
\code{\link{delaytime_gamma}()},
\code{\link{model2desolvefunction}()}
}
\concept{conversions}
