% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movefracpopsize_byname.R
\name{movefracpopsize_byname}
\alias{movefracpopsize_byname}
\title{Move fraction of popsize for named states}
\usage{
movefracpopsize_byname(
  tblpopsize,
  fracpopsize,
  namebefore,
  nameafter,
  samplerandomly
)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize column}

\item{fracpopsize}{fraction of population to move between states}

\item{namebefore}{updatedstate value(s) in tblupdatedstate to remove popsize}

\item{nameafter}{updatedstate value in tblupdatedstate to add popsize}

\item{samplerandomly}{logical specifying whether population moved is
determined by binomial sampling (TRUE and integer populations) or by the
fractional amount (FALSE) with rounding when integer populations.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Move fraction of populations between some number of named updatedstates to
another single named state.
}
