% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedlist2tibblerow.R
\name{namedlist2tibblerow}
\alias{namedlist2tibblerow}
\title{Create 1 row tibble from list where input
values are treated as lists}
\usage{
namedlist2tibblerow(namedlist)
}
\arguments{
\item{namedlist}{named list}
}
\value{
1 row tibble with names as columns
and values as lists
}
\description{
Create 1 row tibble from list where input
values are treated as lists
}
