% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize2probability.R
\name{normalize2probability}
\alias{normalize2probability}
\title{normalize character/numeric vector to sum to 1}
\usage{
normalize2probability(inputvector)
}
\arguments{
\item{inputvector}{character/numeric vector}
}
\value{
normalized vector same type as input.
}
\description{
normalize character/numeric vector to sum to 1
}
