% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_core_equns.R
\name{odin_core_eqns}
\alias{odin_core_eqns}
\title{Create core equations for odin model}
\usage{
odin_core_eqns(compiled_model, states_vec)
}
\arguments{
\item{compiled_model}{compiled model object
created by compModels function compilemodel()}

\item{states_vec}{vector of states in model as character strings
example: base_states <- c("S", "E", "I", "R")}
}
\value{
vector of strings formatted to describe core equation
}
\description{
Core equations describes the movements between states in the model
}
\details{
Format needs to be: update(state) <- state - n_exit + n_enter
}
