% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_draws.R
\name{odin_draws}
\alias{odin_draws}
\title{Draws distributions for numbers changing between compartments/states
Refer to https://mrc-ide.github.io/odin.dust/articles/sir_models.html
for format examples}
\usage{
odin_draws(compiled_model, states_vec)
}
\arguments{
\item{compiled_model}{compiled model object
created by compModels function compilemodel()}

\item{states_vec}{vector of states in model as character strings
example: base_states <- c("S", "E", "I", "R")}
}
\value{
vector of strings formatted to describe draws of numbers
changing between states
}
\description{
In progress: need to add functionality for multinomial draws
}
