% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_probs.R
\name{odin_probs}
\alias{odin_probs}
\title{Calculate individual transition probabilities for odin model
Refer to https://mrc-ide.github.io/odin.dust/articles/sir_models.html
for format examples}
\usage{
odin_probs(compiled_model)
}
\arguments{
\item{compiled_model}{compiled model object
created by compModels function compilemodel()}
}
\value{
vector of strings formatted to describe transition probabilities
}
\description{
Calculate individual transition probabilities for odin model
Refer to https://mrc-ide.github.io/odin.dust/articles/sir_models.html
for format examples
}
