% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peterrow2dxdt.R
\name{peterrow2dxdt}
\alias{peterrow2dxdt}
\title{Get instantaneous change to population}
\usage{
peterrow2dxdt(currrow, currprocessrates)
}
\arguments{
\item{currrow}{Row of petersen matrix with integer values}

\item{currprocessrates}{character vector of process
rates to be added together}
}
\value{
String specifying total instaneous change
across the row (ie for a state)
}
\description{
For a given row of petersen matrix,
add/subtract process
rates as scaled by value in row. Useful to
generate input for deSolve
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
