% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removepopsize_byname.R
\name{removepopsize_byname}
\alias{removepopsize_byname}
\title{Remove popsize for named states}
\usage{
removepopsize_byname(tblpopsize, namevec)
}
\arguments{
\item{tblpopsize}{compiled updatedstate table with popsize value}

\item{namevec}{named list/vector vector specifying updatedstate popsize
values to set. Does not impact unnamed states.}
}
\value{
tibble with columns of updated state names
(updatedstates) and current conditions (popsize)
}
\description{
Remove popsize for named states
}
