% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceglobal.R
\name{replaceglobal}
\alias{replaceglobal}
\title{Replace empty strings in a dataframe column with a given vector}
\usage{
replaceglobal(currtbl, replacecol, replacevec)
}
\arguments{
\item{currtbl}{A dataframe/tibble with columns}

\item{replacecol}{tibble name to replace "" values}

\item{replacevec}{character vector to replace with}
}
\value{
tibble with specified column values ""
replaced with list of the vector to facilitate unnesting.
}
\description{
Replace empty strings in a dataframe column with a given vector
}
