% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleprocesstblbygroups.R
\name{scaleprocesstblbygroups}
\alias{scaleprocesstblbygroups}
\title{Scales processes by grouping}
\usage{
scaleprocesstblbygroups(tblprocess, tblgroups, scalecol)
}
\arguments{
\item{tblprocess}{table of processes to be scaled as generated by compiler}

\item{tblgroups}{table of grouping process for scaling}

\item{scalecol}{column name to scale processes by type (e.g., transition,
migration, interaction)}
}
\value{
named vector with immediate names
}
\description{
For use in compiler
}
