% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleprocesstblbyspace.R
\name{scaleprocesstblbyspace}
\alias{scaleprocesstblbyspace}
\title{Scales processes by space}
\usage{
scaleprocesstblbyspace(tblprocess, tblspace, scalecol)
}
\arguments{
\item{tblprocess}{table of processes to be scaled as generated by compiler}

\item{tblspace}{table of how processes scale between metapopulations}

\item{scalecol}{column name to scale processes by type (e.g., scaletranstion)}
}
\value{
named vector with immediate names
}
\description{
For use in compiler
}
