% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_x0_byname.R
\name{set_x0_byname}
\alias{set_x0_byname}
\title{Generate initial conditions specifying value
for state names}
\usage{
set_x0_byname(tblx0, inputvec)
}
\arguments{
\item{tblx0}{tibble with "updatedstate" column
of state names}

\item{inputvec}{named list with state names as names and
values for population abundances}
}
\value{
Same tibble with "X0" column added with values
}
\description{
Generate initial conditions specifying value
for state names
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
