% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_final_size.R
\name{sir_final_size}
\alias{sir_final_size}
\title{Calculates final size of deterministic SIR model}
\usage{
sir_final_size(r_infinity, r_0)
}
\arguments{
\item{r_infinity}{Proportion (between 0 and 1) of recovered individuals
as time goes toward infinity}

\item{r_0}{Reproductive number}
}
\value{
single value vector (typeL double) containing the final outbreak
size, eg proportion of recovered individuals of the total population
}
\description{
Function containing the final size relation equation for a
basic SIR model for use with root finding tools. Equation from JC Miller
Bull Math Biol (2012) 74:2126. This works for the case when the total
population is N = 1.
}
\details{
Note: Due to equation being for N = 1. \code{\link[=sir_deriv]{sir_deriv()}} uses
explicit frequency dependent transmission on the flow from susceptibles to
infecteds.
}
\examples{
\dontrun{
sir_final_size(1, 3)
}
}
