% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitnamedlist.R
\name{splitnamedlist}
\alias{splitnamedlist}
\title{Slices named list of vectors by combining vector
elements by index in separate named lists}
\usage{
splitnamedlist(namedlist)
}
\arguments{
\item{namedlist}{named list of vectors}
}
\value{
list of named lists of 1-element vectors
}
\description{
Vectors must be same length, use matchlength.R to coerce.
}
