% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitratebynames.R
\name{splitratebynames}
\alias{splitratebynames}
\title{Replaces substrings in string}
\usage{
splitratebynames(currrate, oldnames, newnames)
}
\arguments{
\item{currrate}{String to replace substrings}

\item{oldnames}{substrings to replace (commonly basestates)}

\item{newnames}{Replacing substrings in same order
as oldnames (commonly updated state names)}
}
\value{
Full string with replaced substrings
}
\description{
Useful when updating the basestate in a
rate with the actual state name
}
