% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2petercolumn.R
\name{states2petercolumn}
\alias{states2petercolumn}
\title{Generate column for petersen matrix for a process}
\usage{
states2petercolumn(statesdown, statesup, statelist)
}
\arguments{
\item{statesdown}{State names to decrease in value
(by one). Repeat values to decrease by more than one}

\item{statesup}{State names to increase in value
(by one). Repeat values to increase by more than one}

\item{statelist}{character vector of state names
that specifies order in output matrix}
}
\value{
Column vector (as a matrix) with length(statelist) rows and
values the total integer changes to population
}
\description{
Counts occurrences of values in input vectors and combine to get total
increase or decrease of a population due to a process
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()},
\code{\link{tokenizerate}()}
}
\concept{internal}
