% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2petercolumn.R
\name{states2petercolumn}
\alias{states2petercolumn}
\title{Generate column for petersen matrix for a process}
\usage{
states2petercolumn(statesdown, statesup, statelist)
}
\arguments{
\item{statesdown}{State names to decrease in value
(by one)}

\item{statesup}{statesup State names to increase
in value (by one)}

\item{statelist}{character vector of state names
that specifies order in output matrix}
}
\value{
Column matrix with length(statelist) rows and
values the total integer changes to population.
}
\description{
Input up/down states can be repeated to
increase populations
}
