% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumvector.R
\name{sumvector}
\alias{sumvector}
\title{Sum vector according to type}
\usage{
sumvector(inputvector)
}
\arguments{
\item{inputvector}{numeric or character vector}
}
\value{
1 element numeric or character vector respecting input type
}
\description{
Useful to normalize rates by total
population size when
states are input as numeric or character
}
