% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time2rate.R
\name{time2rate}
\alias{time2rate}
\title{Convert rate to time}
\usage{
time2rate(inputtimes)
}
\arguments{
\item{inputtimes}{numeric or character vector}
}
\value{
vector of same type as input with inverse values
}
\description{
Outputs inverse of input (e.g., 1/x) in same type as input
}
