% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizerate.R
\name{tokenizerate}
\alias{tokenizerate}
\title{Used to isolate user input state and parameter
names within mathematical functions}
\usage{
tokenizerate(ratestr)
}
\arguments{
\item{ratestr}{Character string}
}
\value{
Character vector where each element
are the substrings separated by common
mathematical operations.
}
\description{
Used to isolate user input state and parameter
names within mathematical functions
}
\seealso{
Other internal:
\code{\link{apply_growth_rate_calc}()},
\code{\link{check_inputpopsize}()},
\code{\link{cleanscale}()},
\code{\link{empty2states}()},
\code{\link{fixinput_chains}()},
\code{\link{fixinput_groupnames}()},
\code{\link{grabfromtbl}()},
\code{\link{groupcombinations}()},
\code{\link{grouptypecombinations}()},
\code{\link{intlogic}()},
\code{\link{makenewvariablenamesfromrow}()},
\code{\link{matchlength}()},
\code{\link{matchlength_namedlist}()},
\code{\link{mixedunlist}()},
\code{\link{namedlist2tibblerow}()},
\code{\link{normalize2mean1}()},
\code{\link{normalize2probability}()},
\code{\link{odegenerator}()},
\code{\link{peterrow2dxdt}()},
\code{\link{remove1star}()},
\code{\link{replaceglobal}()},
\code{\link{scaleprocesstblbygroups}()},
\code{\link{scaleprocesstblbyspace}()},
\code{\link{set_x0_byname}()},
\code{\link{split2typecols}()},
\code{\link{splitnamedlist}()},
\code{\link{splitratebynames}()},
\code{\link{states2petercolumn}()},
\code{\link{sumvector}()},
\code{\link{time2rate}()}
}
\concept{internal}
