% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gen_determ_input.R
\name{validate_gen_determ_input}
\alias{validate_gen_determ_input}
\title{Checks validity of input to generalized compartmental model}
\usage{
validate_gen_determ_input(
  init_vals,
  times,
  comp_names,
  trans_matrix,
  subgroups_list,
  intervention_start_time,
  intervention_end_time,
  modifier
)
}
\arguments{
\item{init_vals}{the starting values for populations in each compartment}

\item{times}{vector of time steps to run the model over}

\item{comp_names}{names of the compartments}

\item{trans_matrix}{matrix of transition rates between each of the
compartments}

\item{subgroups_list}{list of one or more vectors of subgroup names to be
used for stratification}

\item{intervention_start_time}{time at which intervention will start}

\item{intervention_end_time}{time at which intervention will end}

\item{modifier}{single value, matrix, or function to be applied to transition
matrix during an intervention}
}
\value{
stops with information or indicates Input Checks Passed
}
\description{
Helper function that runs checks on generalized compartmental model input
parameters, times, and initial conditions
}
